% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_mobility}
\alias{apple_mobility}
\title{Apple Mobility Data}
\format{
A data frame with 2,254,515 rows and 7 variables:
\describe{
\item{\code{country}}{character Country name (not provided for all countries)}
\item{\code{sub_region}}{character Subregion names}
\item{\code{subregion_and_city}}{character Subregion and city names}
\item{\code{geo_type}}{character Type geographical unit. Values: city, country/region, sub-region}
\item{\code{transportation_type}}{character Mode of transport. Values: driving, transit, or walking}
\item{\code{date}}{double Date in yyyy-mm-dd format}
\item{\code{score}}{double Activity score. Indexed to 100 on the first date of observation for a given mode of transport.}
}
}
\source{
https://www.apple.com/covid19/mobility
}
\usage{
apple_mobility
}
\description{
Data from Apple Maps on relative changes in mobility in various cities and countries.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab apple_mobility \cr
   Number of rows \tab 2254515 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-13 \tab 2022-04-12 \tab 2021-02-26 \tab 819 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   country \tab 0 \tab 1 \tab 5 \tab 20 \tab 0 \tab 63 \tab 0 \cr
   sub_region \tab 0 \tab 1 \tab 4 \tab 46 \tab 0 \tab 606 \tab 0 \cr
   subregion_and_city \tab 0 \tab 1 \tab 4 \tab 46 \tab 0 \tab 853 \tab 0 \cr
   geo_type \tab 0 \tab 1 \tab 4 \tab 14 \tab 0 \tab 3 \tab 0 \cr
   transportation_type \tab 0 \tab 1 \tab 7 \tab 7 \tab 0 \tab 3 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   score \tab 608041 \tab 0.73 \tab 122.59 \tab 66.81 \tab 2.43 \tab 83.79 \tab 113.72 \tab 148.8 \tab 2228.83 \tab ▇▁▁▁▁ \cr
}


Data made available by Apple, Inc. at \url{https://www.apple.com/covid19/mobility}, showing relative volume of directions requests per country/region or city compared to a baseline volume on January 13th, 2020. Apple defines the day as midnight-to-midnight, Pacific time. Cities represent usage in greater metropolitan areas and are stably defined during this period. In many countries/regions and cities, relative volume has increased since January 13th, consistent with normal, seasonal usage of Apple Maps. Day of week effects are important to normalize as you use this data. Data that is sent from users’ devices to the Apple Maps service is associated with random, rotating identifiers so Apple does not have a profile of individual movements and searches. Apple Maps has no demographic information about its users, and so cannot make any statements about the representativeness of its usage against the overall population.
}
\references{
See https://www.apple.com/covid19/mobility for detailed terms of use.
}
\author{
Kieran Healy
}
\keyword{datasets}
