% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_deaths_by_sex}
\alias{cdc_deaths_by_sex}
\title{CDC provisional death counts by sex}
\format{
A data frame with 3 rows and 10 variables:
\describe{
\item{\code{data_as_of}}{date Date most recently updated}
\item{\code{sex}}{character Sex}
\item{\code{start_week}}{date Beginning week}
\item{\code{end_week}}{date Ending week}
\item{\code{covid_deaths}}{integer COVID deaths}
\item{\code{total_deaths}}{integer Total deaths}
\item{\code{percent_expected_deaths}}{double COLUMN_DESCRIPTION}
\item{\code{pneumonia_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{pneumonia_and_covid_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{all_influenza_deaths_j09_j11}}{integer COLUMN_DESCRIPTION}
}
}
\source{
Courtesy of Bob Rudis's cdccovidview package
}
\usage{
cdc_deaths_by_sex
}
\description{
Provisional Death Counts for Coronavirus Disease (COVID-19)
}
\details{
c("character", "Date", "Date", "Date", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric")
c("sex", "data_as_of", "start_week", "end_week", "covid_deaths", "total_deaths", "percent_expected_deaths", "pneumonia_deaths", "pneumonia_and_covid_deaths", "all_influenza_deaths_j09_j11")
c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1)
c(4, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(7, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(0, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(3, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(0, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(NA, 18382, 18293, 18377, NA, NA, NA, NA, NA, NA)
c(NA, 18382, 18293, 18377, NA, NA, NA, NA, NA, NA)
c(NA, 18382, 18293, 18377, NA, NA, NA, NA, NA, NA)
c(NA, 1, 1, 1, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, NA, 11507.3333333333, 237795, 0.97, 20908.3333333333, 5100.33333333333, 1940.33333333333)
c(NA, NA, NA, NA, 10231.3977702625, 206241.057260673, 0, 18248.3963496339, 4559.67381874332, 1682.21054964393)
c(NA, NA, NA, NA, 1, 25, 0.97, 1, 1, 0)
c(NA, NA, NA, NA, 7470.5, 172555, 0.97, 14545, 3258, 1416)
c(NA, NA, NA, NA, 14940, 345085, 0.97, 29089, 6515, 2832)
c(NA, NA, NA, NA, 17260.5, 356680, 0.97, 31362, 7650, 2910.5)
c(NA, NA, NA, NA, 19581, 368275, 0.97, 33635, 8785, 2989)
c(NA, NA, NA, NA, "▇▁▁▇▇", "▃▁▁▁▇", "▁▁▇▁▁", "▃▁▁▁▇", "▇▁▁▇▇", "▃▁▁▁▇")

The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}). Please see the indicated reference for all the caveats and precise meanings for each field.
}
\references{
\url{https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true}
}
\author{
Kieran Healy
}
\keyword{datasets}
