% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_deaths_by_state}
\alias{cdc_deaths_by_state}
\title{CDC provisional death counts by state}
\format{
A data frame with 54 rows and 13 variables:
\describe{
\item{\code{state}}{character COLUMN_DESCRIPTION}
\item{\code{covid_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{total_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{percent_expected_deaths}}{double COLUMN_DESCRIPTION}
\item{\code{pneumonia_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{pneumonia_and_covid_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{all_influenza_deaths_j09_j11}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
cdc_deaths_by_state
}
\description{
CDC Surveillance Network provisional death counts
}
\details{
The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET. Please see the indicated reference for all the caveats and precise meanings for each field.
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}).
}
\references{
\url{https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true}
}
\author{
Kieran Healy
}
\keyword{datasets}
