% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_deaths_by_week}
\alias{cdc_deaths_by_week}
\title{CDC Provisional death counts by week}
\format{
A data frame with 13 rows and 10 variables:
\describe{
\item{\code{data_as_of}}{date When the data were most recently recorded}
\item{\code{start_week}}{date Start week}
\item{\code{end_week}}{double End week}
\item{\code{covid_deaths}}{integer COVID deaths}
\item{\code{total_deaths}}{integer Total deaths}
\item{\code{percent_expected_deaths}}{double COLUMN_DESCRIPTION}
\item{\code{pneumonia_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{pneumonia_and_covid_deaths}}{integer COLUMN_DESCRIPTION}
\item{\code{all_influenza_deaths_j09_j11}}{integer COLUMN_DESCRIPTION}
\item{\code{pneumonia_influenza_and_covid_19_deaths}}{integer COLUMN_DESCRIPTION}
}
}
\source{
Courtesy of Bob Rudis's cdccovidview package
}
\usage{
cdc_deaths_by_week
}
\description{
Provisional Death Counts for Coronavirus Disease (COVID-19)
}
\details{
c("Date", "Date", "Date", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric")
c("data_as_of", "start_week", "end_week", "covid_deaths", "total_deaths", "percent_expected_deaths", "pneumonia_deaths", "pneumonia_and_covid_deaths", "all_influenza_deaths_j09_j11", "pneumonia_influenza_and_covid_19_deaths")
c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1)
c(18382, 18293, 18293, NA, NA, NA, NA, NA, NA, NA)
c(18382, 18377, 18377, NA, NA, NA, NA, NA, NA, NA)
c(18382, 18335, 18335, NA, NA, NA, NA, NA, NA, NA)
c(1, 13, 13, NA, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, 2655.46153846154, 54875.8461538462, 0.965384615384615, 4825, 1177, 447.769230769231, 6690.23076923077)
c(NA, NA, NA, 4194.37332656073, 9864.46311975597, 0.17222153294596, 2217.1934962921, 1863.76286045194, 156.185335337089, 4292.62196786234)
c(NA, NA, NA, 0, 24387, 0.45, 2219, 0, 58, 3553)
c(NA, NA, NA, 0, 53940, 0.97, 3671, 0, 427, 4165)
c(NA, NA, NA, 49, 56831, 0.97, 3692, 25, 494, 4275)
c(NA, NA, NA, 2659, 57299, 0.99, 5598, 1220, 536, 7397)
c(NA, NA, NA, 11864, 65676, 1.19, 9580, 5281, 619, 16272)
c(NA, NA, NA, "▇▁▁▂▁", "▁▁▁▇▂", "▁▁▁▇▂", "▇▃▁▁▂", "▇▁▁▂▁", "▁▁▁▇▇", "▇▁▁▂▁")

The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}). Please see the indicated reference for all the caveats and precise meanings for each field.
}
\references{
\url{https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true}
}
\author{
Kieran Healy
}
\keyword{datasets}
