% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_hospitalizations}
\alias{cdc_hospitalizations}
\title{CDC Laboratory Confirmed COVID-19-Associated Hospitalization in the US}
\format{
A data frame with 4,590 rows and 8 variables:
\describe{
\item{\code{catchment}}{character COLUMN_DESCRIPTION}
\item{\code{network}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{character COLUMN_DESCRIPTION}
\item{\code{mmwr_year}}{character COLUMN_DESCRIPTION}
\item{\code{mmwr_week}}{character COLUMN_DESCRIPTION}
\item{\code{age_category}}{character COLUMN_DESCRIPTION}
\item{\code{cumulative_rate}}{double COLUMN_DESCRIPTION}
\item{\code{weekly_rate}}{double COLUMN_DESCRIPTION}
}
}
\source{
Courtesy of Bob Rudis's cdccovidview package
}
\usage{
cdc_hospitalizations
}
\description{
Courtesy of Bob Rudis's cdccovidview package
}
\details{
c("character", "character", "character", "character", "character", "character", "numeric", "numeric")
c("catchment", "network", "year", "mmwr_year", "mmwr_week", "age_category", "cumulative_rate", "weekly_rate")
c(0, 0, 0, 0, 0, 0, 1836, 1836)
c(1, 1, 1, 1, 1, 1, 0.6, 0.6)
c(4, 3, 4, 4, 2, 3, NA, NA)
c(14, 9, 4, 4, 2, 8, NA, NA)
c(0, 0, 0, 0, 0, 0, NA, NA)
c(15, 3, 1, 1, 30, 9, NA, NA)
c(0, 0, 0, 0, 0, 0, NA, NA)
c(NA, NA, NA, NA, NA, NA, 108.540777051561, 10.8326797385621)
c(NA, NA, NA, NA, NA, NA, 177.971385145269, 19.4395350884378)
c(NA, NA, NA, NA, NA, NA, 0, 0)
c(NA, NA, NA, NA, NA, NA, 2.9, 0.5)
c(NA, NA, NA, NA, NA, NA, 38.85, 4.2)
c(NA, NA, NA, NA, NA, NA, 138.875, 12.675)
c(NA, NA, NA, NA, NA, NA, 1684.1, 269.2)
c(NA, NA, NA, NA, NA, NA, "▇▁▁▁▁", "▇▁▁▁▁")

The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}).
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html}
}
\author{
Kieran Healy
}
\keyword{datasets}
