% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country Names and ISO codes}
\format{
A data frame with 215 rows and 4 variables:
\describe{
\item{\code{cname}}{character Country name}
\item{\code{iso3}}{character ISO 3 designation}
\item{\code{iso2}}{character ISO 2 designation}
\item{\code{continent}}{Continent}
}
}
\usage{
countries
}
\description{
Convenience table of country names and their abbreviated names
}
\details{
Produced from the ECDC tables in the covdata package
}
\references{
ISO 2: \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2} ISO 3: \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}
}
\author{
Kieran Healy
}
\keyword{datasets}
