% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covda-package.R
\docType{package}
\name{covda-package}
\alias{covda}
\alias{covda-package}
\title{covda: Analyze Cases With Countries In Cov19}
\description{
Analyze Cases With Countries In Cov19.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-AntiRosa}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-AntiRosa/issues}
}

}
\author{
\strong{Maintainer}: Linling Bai \email{353621303@qq.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
