% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat_daily}
\alias{covnat_daily}
\title{Daily international COVID-19 cases and deaths for 2020}
\format{
A tibble with 61,836 rows and 8 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{cname}{Name of country (character)}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this day}
\item{deaths}{N reported COVID-19 deaths for this day}
\item{pop}{Country population from Eurostat or UN data}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this day}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this day}
}
}
\source{
\url{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}
}
\usage{
covnat_daily
}
\description{
A dataset containing daily national-level ECDC data on COVID-19. Archived as of December 14th 2020.
ECDC switched to a weekly reporting schedule for the COVID-19 situation worldwide and in the EU/EEA and the UK on 17 December 2020. Daily updates have been discontinued from 14 December 2020.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat_dai... \cr
   Number of rows \tab 61836 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 2 \cr
   Date \tab 1 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   cname \tab 0 \tab 1 \tab 4 \tab 42 \tab 0 \tab 213 \tab 0 \cr
   iso3 \tab 0 \tab 1 \tab 3 \tab 3 \tab 0 \tab 213 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-31 \tab 2020-12-14 \tab 2020-07-21 \tab 350 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 1156.33 \tab 6782.63 \tab -8261 \tab 0 \tab 15 \tab 275.00 \tab 234633 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 26.08 \tab 131.29 \tab -1918 \tab 0 \tab 0 \tab 4.00 \tab 4928 \tab ▁▇▁▁▁ \cr
   pop \tab 59 \tab 1 \tab 40987698.23 \tab 153129379.34 \tab 815 \tab 1293120 \tab 7169456 \tab 28515829.00 \tab 1433783692 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1 \tab 100686.99 \tab 607743.06 \tab 0 \tab 129 \tab 2055 \tab 24650.00 \tab 16256754 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1 \tab 3104.89 \tab 15545.84 \tab 0 \tab 1 \tab 42 \tab 464.25 \tab 299177 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
