% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat_weekly}
\alias{covnat_weekly}
\title{Weekly International COVID-19 cases and deaths, current as of Wednesday, March 24, 2021}
\format{
A tibble with 12,086 rows and 11 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{year_week}{Year and week of reporting (character, YYYY-WW)}
\item{cname}{Name of country (character)}
\item{pop}{Country population from Eurostat or UN data}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this week}
\item{deaths}{N reported COVID-19 deaths for this week}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this week}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this week}
\item{r14_cases}{14-day notification rate of reported COVID-19 cases per 100,000 population}
\item{r14_deaths}{14-day notification rate of reported COVID-19 cases per 100,000 population}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat_weekly
}
\description{
A dataset containing weekly national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat_wee... \cr
   Number of rows \tab 12086 \cr
   Number of columns \tab 11 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 3 \cr
   numeric \tab 7 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-30 \tab 2021-03-08 \tab 2020-08-31 \tab 63 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   year_week \tab 0 \tab 1.00 \tab 7 \tab 7 \tab 0 \tab 63 \tab 0 \cr
   cname \tab 0 \tab 1.00 \tab 4 \tab 33 \tab 0 \tab 220 \tab 0 \cr
   iso3 \tab 378 \tab 0.97 \tab 3 \tab 3 \tab 0 \tab 214 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pop \tab 0 \tab 1.00 \tab 81567830.14 \tab 376867865.28 \tab 809.00 \tab 1318442.00 \tab 8606033.00 \tab 32365998.00 \tab 4.498460e+09 \tab ▇▁▁▁▁ \cr
   cases \tab 0 \tab 1.00 \tab 21902.32 \tab 128671.73 \tab -28304.00 \tab 12.00 \tab 252.00 \tab 3522.00 \tab 2.699838e+06 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1.00 \tab 487.91 \tab 2694.67 \tab -875.00 \tab 0.00 \tab 3.00 \tab 55.75 \tab 4.827700e+04 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1.00 \tab 402248.05 \tab 2559277.73 \tab 0.00 \tab 275.25 \tab 4672.50 \tab 60645.00 \tab 5.349725e+07 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1.00 \tab 10361.51 \tab 60006.64 \tab 0.00 \tab 3.00 \tab 79.00 \tab 1098.75 \tab 1.282028e+06 \tab ▇▁▁▁▁ \cr
   r14_cases \tab 220 \tab 0.98 \tab 92.10 \tab 204.41 \tab -130.01 \tab 1.08 \tab 10.44 \tab 85.15 \tab 4.345370e+03 \tab ▇▁▁▁▁ \cr
   r14_deaths \tab 220 \tab 0.98 \tab 16.15 \tab 41.56 \tab -67.13 \tab 0.00 \tab 1.05 \tab 12.39 \tab 1.395030e+03 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
