% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus}
\alias{covus}
\title{COVID-19 data for the USA, current as of Tuesday, June 16, 2020}
\format{
A tibble with 85,935 rows and 7 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{state}{Two letter State abbreviation (character)}
\item{fips}{State FIPS code (character)}
\item{\code{data_quality_grade}}{character Data quality as assessed by COVID Tracking Project staff}
\item{measure}{Outcome measure for this date}
\item{count}{Count of measure}
\item{\code{measure_label}}{character Outcome measure, suitable for use as a plot label}
}
}
\source{
The COVID-19 Tracking Project \url{https://covidtracking.com}
}
\usage{
covus
}
\description{
A dataset containing US state-level data on COVID-19
}
\details{
The measures tracked by the COVID tracking project are as follows:
|measure                  |measure_label                                          |
|:------------------------|:------------------------------------------------------|
|positive                 |Positive Tests                                         |
|negative                 |Negative Tests                                         |
|pending                  |Pending Tests                                          |
|hospitalized_currently   |Currently Hospitalized                                 |
|hospitalized_cumulative  |Cumulative Hospitalized                                |
|in_icu_currently         |Currently in ICU                                       |
|in_icu_cumulative        |Cumulative in ICU                                      |
|on_ventilator_currently  |Currently on Ventilator                                |
|on_ventilator_cumulative |Cumulative on Ventilator                               |
|recovered                |Recovered                                              |
|death                    |Deaths                                                 |
|total_tests_viral        |Total number of PCR tests performed                    |
|positive_tests_viral     |Total number of positive PCR tests                     |
|negative_tests_viral     |Total number of negative PCR tests                     |
|positive_cases_viral     |Total number of positive cases measured with PCR tests |
Not all measures are reported by all states.
}
\keyword{datasets}
