% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fmt_nc}
\alias{fmt_nc}
\title{fmt_nc}
\usage{
fmt_nc(x)
}
\arguments{
\item{x}{df}
}
\value{
formatted string
}
\description{
Format fmt_nc in df
}
\details{
use in fn documentation
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Kieran Healy
}
