% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fmt_nr}
\alias{fmt_nr}
\title{fmt_nr}
\usage{
fmt_nr(x)
}
\arguments{
\item{x}{df}
}
\value{
formatted string
}
\description{
Format fmt_nr in df
}
\details{
use in fn documentation
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Kieran Healy
}
