% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Analyze Cases With Countries In Cov19
This data include 379 countries in the world}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 143825 rows and 5 columns.
}
\source{
https://github.com/RamiKrispin/coronavirus
The coronavirus package provides a tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic. The raw data pulled from the Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus repository.
}
\usage{
key_crop_yields
}
\description{
Analyze Cases With Countries In Cov19
This data include 379 countries in the world
}
\keyword{datasets}
