% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mmwr_week_to_date}
\alias{mmwr_week_to_date}
\title{FUNCTION_TITLE}
\source{
http://
}
\usage{
mmwr_week_to_date(year, week, day = NULL)
}
\arguments{
\item{year}{PARAM_DESCRIPTION}

\item{week}{PARAM_DESCRIPTION}

\item{day}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MMWRweek]{MMWRweek2Date}}
}
\author{
Kieran Healy
}
