% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MMWRweek2Date}
\alias{MMWRweek2Date}
\title{FUNCTION_TITLE}
\source{
http://
}
\usage{
MMWRweek2Date(MMWRyear, MMWRweek, MMWRday = NULL)
}
\arguments{
\item{MMWRyear}{PARAM_DESCRIPTION}

\item{MMWRweek}{PARAM_DESCRIPTION}

\item{MMWRday}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{

}
\author{
Kieran Healy
}
