% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MMWRweekday}
\alias{MMWRweekday}
\title{FUNCTION_TITLE}
\source{
http://
}
\usage{
MMWRweekday(date)
}
\arguments{
\item{date}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\references{

}
\author{

}
