% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_pud}
\alias{nchs_pud}
\title{COVID-19 Case Surveillance Public Use Data}
\format{
A tibble with 1,748,965 rows and 11 variables:
\describe{
\item{\code{cdc_report_dt}}{date Initial case report date to CDC}
\item{\code{pos_spec_dt}}{date Date of first positive specimen collection}
\item{\code{onset_dt}}{date Symptom onset date, if symptomatic}
\item{\code{current_status}}{character Case Status: Laboratory-confirmed case or Probable case}
\item{\code{sex}}{character Sex: Male; Female; Unknown; Other}
\item{\code{age_group}}{character Age Group: 0-9 Years; 10-19 Years; 20-39 Years; 40-49 Years; 50-59 Years; 60-69 Years; 70-79 Years; 80+ Years}
\item{\code{race_ethnicity}}{character Race and ethnicity (combined): Hispanic/Latino; American Indian / Alaska Native, Non-Hispanic; Asian, Non-Hispanic; Black, Non-Hispanic; Native Hawaiian / Other Pacific Islander, Non-Hispanic; White, Non-Hispanic; Multiple/Other, Non-Hispanic}
\item{\code{hosp_yn}}{character Hospitalization status}
\item{\code{icu_yn}}{character ICU admission status}
\item{\code{death_yn}}{character Death status}
\item{\code{medcond_yn}}{character Presence of underlying comorbidity or disease}
}
}
\source{
https://data.cdc.gov/Case-Surveillance/COVID-19-Case-Surveillance-Public-Use-Data/vbim-akqf
}
\usage{
nchs_pud
}
\description{
Deidentified patient-level data reported to U.S. states and autonomous reporting entities, including New York City and the District of Columbia (D.C.), as well as U.S. territories and states.
}
\details{
c("character", "character", "character", "character", "character", "character", "character", "character", "Date", "Date", "Date")
c("current_status", "sex", "age_group", "race_ethnicity", "hosp_yn", "icu_yn", "death_yn", "medcond_yn", "cdc_report_dt", "pos_spec_dt", "onset_dt")
c(0, 15, 38, 9, 0, 0, 0, 0, 0, 1052553, 969687)
c(1, 0.999991423499041, 0.999978272864237, 0.999994854099425, 1, 1, 1, 1, 1, 0.398185212397046, 0.445565234295712)
c(8, 4, 6, 7, 2, 2, 2, 2, NA, NA, NA)
c(9, 7, 8, 28, 7, 7, 7, 7, NA, NA, NA)
c(0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA)
c(2, 4, 10, 8, 3, 3, 3, 3, NA, NA, NA)
c(0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, 18262, 18265, 18262)
c(NA, NA, NA, NA, NA, NA, NA, NA, 18424, 18437, 18439)
c(NA, NA, NA, NA, NA, NA, NA, NA, 18383, 18370, 18368)
c(NA, NA, NA, NA, NA, NA, NA, NA, 116, 137, 178)

Each row is a deidentified patient. These deidentified data include demographic characteristics, exposure history, disease severity indicators and outcomes, clinical data, laboratory diagnostic test results, and comorbidities. All data elements can be found on the COVID-19 case report form located at \code{www.cdc.gov/coronavirus/2019-ncov/downloads/pui-form.pdf}.
The COVID-19 case surveillance data are dynamic; case reports can be modified at any time by the reporting jurisdiction as new information becomes available (i.e., data are subject to change). Furthermore, reporting jurisdictions may report cases late. Version updates to the detailed and limited datasets will be available for request once a month.
The datasets will include all cases with an initial report date of case to CDC at least 30 days prior to the creation of the previously updated datasets. This month lag will allow case reporting to be stabilized and ensure that time-dependent outcome data, including death, are accurately captured.
Questions that have been left unanswered (blank) on the case report form are re-classified to an Unknown value, if applicable to the question. For example, in the question “Was the patient hospitalized?”, where the possible answer choices include “Yes”, “No”, or “Unknown”, the missing value is re-coded to the Unknown answer option if the respondent did not answer the question.
The initial report date of the case to CDC is intended to be completed by the reporting jurisdiction when data are submitted. If blank, this variable is completed using the date the data file was first submitted to CDC.
To prevent release of data that could be used to identify persons, data cells are suppressed for low frequency (<5) records and indirect identifiers (date of first positive specimen). Suppression includes uncommon combinations of demographic characteristics (sex, age group, race/ethnicity). Suppressed values are re-coded to the NA answer option.
}
\author{
Kieran Healy
}
\keyword{datasets}
