% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nytcovus}
\alias{nytcovus}
\title{NYT COVID-19 data for the US, current as of Sunday, November  8, 2020}
\format{
A tibble with 292 rows and 3 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{cases}{Cumulative N  reported cases}
\item{deaths}{Cumulative N reported deaths}
}
}
\source{
The New York Times \url{https://github.com/nytimes/covid-19-data}.
For details on the methods and limitations see \url{https://github.com/nytimes/covid-19-data}.
}
\usage{
nytcovus
}
\description{
A dataset containing US national-level data on COVID-19, collected by the New York Times.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nytcovus \cr
   Number of rows \tab 292 \cr
   Number of columns \tab 3 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-21 \tab 2020-11-07 \tab 2020-06-14 \tab 292 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 3183428.6 \tab 3007387.99 \tab 1 \tab 269346.50 \tab 2116516 \tab 5850158.5 \tab 9957051 \tab ▇▂▂▂▂ \cr
   deaths \tab 0 \tab 1 \tab 106388.2 \tab 80797.61 \tab 0 \tab 7584.25 \tab 115992 \tab 179884.5 \tab 237567 \tab ▇▂▅▅▅ \cr
}
}
\keyword{datasets}
