% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nytexcess}
\alias{nytexcess}
\title{NYT Excess Mortality Estimates, current as of Tuesday, October 19, 2021}
\format{
A tibble with 7,258 rows and 12 columns
\describe{
\item{\code{country}}{character Country Name}
\item{\code{placename}}{character Place Name}
\item{\code{frequency}}{character Reporting period. Weekly or monthly, depending on how the data is recorded.}
\item{\code{start_date}}{date The first date included in the period.}
\item{\code{end_date}}{date The last date included in the period,}
\item{\code{year}}{character Year of data. Note that this variable is of type character and not integer because several observations are notes to the effect that the year is an average of two years.}
\item{\code{month}}{integer Numerical month.}
\item{\code{week}}{integer Numerical week.}
\item{\code{deaths}}{integer  The total number of confirmed deaths recorded from any cause.}
\item{\code{expected_deaths}}{integer The baseline number of expected deaths, calculated from a historical average. See details below.}
\item{\code{excess_deaths}}{integer The number of deaths minus the expected deaths.}
\item{\code{baseline}}{character The years used to calculate expected_deaths.}
}
}
\source{
The New York Times \url{https://github.com/nytimes/covid-19-data/tree/master/excess-deaths}.
}
\usage{
nytexcess
}
\description{
All-cause mortality is widely used by demographers and other researchers to understand the full impact of deadly events, including epidemics, wars and natural disasters. The totals in this data include deaths from Covid-19 as well as those from other causes, likely including people who could not be treated or did not seek treatment for other conditions.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nytexcess \cr
   Number of rows \tab 7258 \cr
   Number of columns \tab 12 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 5 \cr
   Date \tab 2 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   country \tab 0 \tab 1.00 \tab 4 \tab 14 \tab 0 \tab 35 \tab 0 \cr
   placename \tab 6883 \tab 0.05 \tab 6 \tab 8 \tab 0 \tab 4 \tab 0 \cr
   frequency \tab 0 \tab 1.00 \tab 6 \tab 7 \tab 0 \tab 2 \tab 0 \cr
   year \tab 0 \tab 1.00 \tab 4 \tab 17 \tab 0 \tab 15 \tab 0 \cr
   baseline \tab 5990 \tab 0.17 \tab 20 \tab 25 \tab 0 \tab 7 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   start_date \tab 768 \tab 0.89 \tab 2010-01-09 \tab 2020-12-23 \tab 2018-02-05 \tab 1267 \cr
   end_date \tab 768 \tab 0.89 \tab 2010-01-15 \tab 2020-12-29 \tab 2018-02-11 \tab 1267 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   month \tab 0 \tab 1.00 \tab 6.60 \tab 3.36 \tab 1 \tab 4.00 \tab 7.0 \tab 9.0 \tab 12 \tab ▇▆▆▆▇ \cr
   week \tab 666 \tab 0.91 \tab 26.77 \tab 14.58 \tab 2 \tab 14.00 \tab 27.0 \tab 39.0 \tab 52 \tab ▇▇▇▇▇ \cr
   deaths \tab 0 \tab 1.00 \tab 7968.24 \tab 14334.14 \tab 455 \tab 1460.00 \tab 2395.5 \tab 10486.0 \tab 141292 \tab ▇▁▁▁▁ \cr
   expected_deaths \tab 5990 \tab 0.17 \tab 9237.09 \tab 15850.00 \tab 548 \tab 1443.00 \tab 2423.0 \tab 10771.5 \tab 139343 \tab ▇▁▁▁▁ \cr
   excess_deaths \tab 5990 \tab 0.17 \tab 1195.43 \tab 3242.72 \tab -6721 \tab -42.25 \tab 76.5 \tab 926.0 \tab 30400 \tab ▇▂▁▁▁ \cr
}


Expected deaths for each area based on historical data for the same time of year. These expected deaths are the basis for our excess death calculations, which estimate how many more people have died this year than in an average year.

The number of years used in the historical averages changes depending on what data is available, whether it is reliable and underlying demographic changes. See Data Sources for the years used to calculate the baselines. The baselines do not adjust for changes in age or other demographics, and they do not account for changes in total population.

The number of expected deaths are not adjusted for how non-Covid-19 deaths may change during the outbreak, which will take some time to figure out. As countries impose control measures, deaths from causes like road accidents and homicides may decline. And people who die from Covid-19 cannot die later from other causes, which may reduce other causes of death. Both of these factors, if they play a role, would lead these baselines to understate, rather than overstate, the number of excess deaths.
}
\references{
For further details on these data see \url{https://github.com/nytimes/covid-19-data/tree/master/excess-deaths}
}
\author{
Kieran Healy
}
\keyword{datasets}
