% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{start_date}
\alias{start_date}
\title{FUNCTION_TITLE}
\source{
http://
}
\usage{
start_date(year)
}
\arguments{
\item{year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
AUTHOR_NAME
}
