% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tabular}
\alias{tabular}
\title{tabular}
\source{
http://
}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Other args}
}
\value{
Rd table
}
\description{
Make an Rd table from a data frame
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Kieran Healy
}
