% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_mobility}
\alias{apple_mobility}
\title{Apple Mobility Data}
\format{
A data frame with 1,402,609 rows and 8 variables:
\describe{
\item{\code{geo_type}}{character Type geographical unit. Values: city or country/region}
\item{\code{region}}{character Name of geographical unit.}
\item{\code{transportation_type}}{character Mode of transport. Values: Driving, Transit, or Walking}
\item{\code{alternative_name}}{character Name of \code{region} in local language}
\item{\code{sub_region}}{character Subregion}
\item{\code{country}}{character Country name (not provided for all countries)}
\item{\code{date}}{double Date in yyyy-mm-dddd format}
\item{\code{score}}{double Activity score. Indexed to 100 on the first date of observation for a given mode of transport.}
}
}
\source{
https://www.apple.com/covid19/mobility
}
\usage{
apple_mobility
}
\description{
Data from Apple Maps on relative changes in mobility in various cities and countries.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab apple_mobility \cr
   Number of rows \tab 1402609 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 6 \cr
   Date \tab 1 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   geo_type \tab 0 \tab 1.00 \tab 4 \tab 14 \tab 0 \tab 4 \tab 0 \cr
   region \tab 0 \tab 1.00 \tab 4 \tab 48 \tab 0 \tab 2325 \tab 0 \cr
   transportation_type \tab 0 \tab 1.00 \tab 7 \tab 7 \tab 0 \tab 3 \tab 0 \cr
   alternative_name \tab 1090453 \tab 0.22 \tab 2 \tab 85 \tab 0 \tab 519 \tab 0 \cr
   sub_region \tab 407238 \tab 0.71 \tab 4 \tab 33 \tab 0 \tab 162 \tab 0 \cr
   country \tab 45747 \tab 0.97 \tab 5 \tab 20 \tab 0 \tab 47 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-13 \tab 2020-11-06 \tab 2020-06-10 \tab 299 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   score \tab 18833 \tab 0.99 \tab 124.54 \tab 67.65 \tab 0.44 \tab 84.79 \tab 115.16 \tab 153.24 \tab 2148.12 \tab ▇▁▁▁▁ \cr
}


Data made available by Apple, Inc. at \url{https://www.apple.com/covid19/mobility}, showing relative volume of directions requests per country/region or city compared to a baseline volume on January 13th, 2020. Apple defines the day as midnight-to-midnight, Pacific time. Cities represent usage in greater metropolitan areas and are stably defined during this period. In many countries/regions and cities, relative volume has increased since January 13th, consistent with normal, seasonal usage of Apple Maps. Day of week effects are important to normalize as you use this data. Data that is sent from users’ devices to the Apple Maps service is associated with random, rotating identifiers so Apple does not have a profile of individual movements and searches. Apple Maps has no demographic information about its users, and so cannot make any statements about the representativeness of its usage against the overall population.
}
\references{
See https://www.apple.com/covid19/mobility for detailed terms of use.
}
\author{
Kieran Healy
}
\keyword{datasets}
