% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_hospitalizations}
\alias{cdc_hospitalizations}
\title{CDC Laboratory Confirmed COVID-19-Associated Hospitalization in the US}
\format{
A data frame with 4,590 rows and 8 variables:
\describe{
\item{\code{catchment}}{character COLUMN_DESCRIPTION}
\item{\code{network}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{character COLUMN_DESCRIPTION}
\item{\code{mmwr_year}}{character COLUMN_DESCRIPTION}
\item{\code{mmwr_week}}{character COLUMN_DESCRIPTION}
\item{\code{age_category}}{character COLUMN_DESCRIPTION}
\item{\code{cumulative_rate}}{double COLUMN_DESCRIPTION}
\item{\code{weekly_rate}}{double COLUMN_DESCRIPTION}
}
}
\source{
Courtesy of Bob Rudis's cdccovidview package
}
\usage{
cdc_hospitalizations
}
\description{
Courtesy of Bob Rudis's cdccovidview package
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab cdc_hospitalizations \cr
   Number of rows \tab 4590 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 6 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   catchment \tab 0 \tab 1 \tab 4 \tab 14 \tab 0 \tab 15 \tab 0 \cr
   network \tab 0 \tab 1 \tab 3 \tab 9 \tab 0 \tab 3 \tab 0 \cr
   year \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 1 \tab 0 \cr
   mmwr_year \tab 0 \tab 1 \tab 4 \tab 4 \tab 0 \tab 1 \tab 0 \cr
   mmwr_week \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 30 \tab 0 \cr
   age_category \tab 0 \tab 1 \tab 3 \tab 8 \tab 0 \tab 9 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cumulative_rate \tab 1683 \tab 0.63 \tab 112.03 \tab 180.69 \tab 0 \tab 3.3 \tab 42.1 \tab 145.5 \tab 1702.7 \tab ▇▁▁▁▁ \cr
   weekly_rate \tab 1683 \tab 0.63 \tab 10.60 \tab 18.83 \tab 0 \tab 0.5 \tab 4.4 \tab 12.3 \tab 271.0 \tab ▇▁▁▁▁ \cr
}


The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}).
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html}
}
\author{
Kieran Healy
}
\keyword{datasets}
