% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_hospitalizations}
\alias{cdc_hospitalizations}
\title{CDC Laboratory Confirmed COVID-19-Associated Hospitalization in the US}
\format{
A data frame with 4590 rows and 8 variables:
\describe{
\item{\code{catchment}}{character COLUMN_DESCRIPTION}
\item{\code{network}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{character COLUMN_DESCRIPTION}
\item{\code{mmwr_year}}{character COLUMN_DESCRIPTION}
\item{\code{mmwr_week}}{character COLUMN_DESCRIPTION}
\item{\code{age_category}}{character COLUMN_DESCRIPTION}
\item{\code{cumulative_rate}}{double COLUMN_DESCRIPTION}
\item{\code{weekly_rate}}{double COLUMN_DESCRIPTION}
}
}
\source{
Courtesy of Bob Rudis's cdccovidview package
}
\usage{
cdc_hospitalizations
}
\description{
Courtesy of Bob Rudis's cdccovidview package
}
\details{
The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}).
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html}
}
\author{
Kieran Healy
}
\keyword{datasets}
