% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coronanet}
\alias{coronanet}
\title{CoronaNet Government Response Project data}
\format{
A tibble with 18,074 rows and 33 variables:
\describe{
\item{\code{record_id}}{character Unique identifier for each unique policy record}
\item{\code{policy_id}}{double Identifier linking new policies with subsequent updates to policies}
\item{\code{recorded_date}}{date-time When the record was entered into the dataset}
\item{\code{date_announced}}{date When the policy was announced}
\item{\code{date_start}}{date When the policy went into effect}
\item{\code{date_end}}{date When the policy ended (if it has an explicit end date)}
\item{\code{entry_type}}{character Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.}
\item{\code{event_description}}{character A short description of the policy change}
\item{\code{domestic_policy}}{double Indicates where policy targets an area within the initiating country (i.e. is domestic in nature)}
\item{\code{type}}{character The category of the policy}
\item{\code{type_sub_cat}}{character The sub-category of the policy (if one exists)}
\item{\code{type_text}}{double Any additional information about the policy type (such as the number of ventilators/days of quarantine/etc.)}
\item{\code{index_high_est}}{double The high (95\% posterior density) estimate of the country policy activity score (0-100)}
\item{\code{index_med_est}}{double The median (most likely) estimate of the country policy activity score (0-100)}
\item{\code{index_low_est}}{double The low (95\% posterior density) estimate of the country policy activity score (0-100)}
\item{\code{index_country_rank}}{double The relative rank by each day for each country on the policy activity score}
\item{\code{country}}{character The country initiating the policy}
\item{\code{init_country_level}}{character Whether the policy came from the national level or a sub-national unit}
\item{\code{province}}{character Name of sub-national unit}
\item{\code{target_country}}{character Which foreign country a policy is targeted at (i.e. travel policies)}
\item{\code{target_geog_level}}{character Whether the target of the policy is a country as a whole or a sub-national unit of that country}
\item{\code{target_region}}{character The name of a regional grouping (like ASEAN) that is a target of the policy (if any)}
\item{\code{target_province}}{character The name of a province targeted by the policy (if any)}
\item{\code{target_city}}{character The name of a city targeted by the policy (if any)}
\item{\code{target_other}}{logical Any geographical entity that does not fit into the targeted categories coded elsewhere}
\item{\code{target_who_what}}{character Who the policy is targeted at}
\item{\code{target_direction}}{character Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)}
\item{\code{travel_mechanism}}{character If a travel policy, what kind of transportation it affects}
\item{\code{compliance}}{character Whether the policy is voluntary or mandatory}
\item{\code{enforcer}}{character What unit in the country is responsible for enforcement}
\item{\code{link}}{character A link to at least one source for the policy}
\item{\code{iso3}}{character 3-digit ISO country code}
\item{\code{iso2}}{character 2-digit ISO country code}
}
}
\source{
\itemize{
\item Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” BETAVersion 1.0. \url{https://www.coronanet-project.org}.
}
}
\usage{
coronanet
}
\description{
Contains variables from the CoronaNet Government Response Project (\verb{https://coronanet-project.org}), representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. See below for citation information.
}
\details{
c("character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "character", "Date", "Date", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric", "POSIXct", "POSIXct")
c("record_id", "policy_id", "entry_type", "event_description", "type", "type_sub_cat", "country", "init_country_level", "province", "target_country", "target_geog_level", "target_region", "target_province", "target_city", "target_other", "target_who_what", "target_direction", "travel_mechanism", "compliance", "enforcer", "link", "iso3", "iso2", "date_start", "date_end", "domestic_policy", "type_text", "index_high_est", "index_med_est", "index_low_est", "index_country_rank", "recorded_date", "date_announced"
)
c(0, 0, 0, 0, 0, 6756, 0, 0, 10610, 186, 11965, 17864, 16672, 17588, 18074, 9539, 14757, 12859, 457, 0, 146, 48, 69, 0, 12926, 0, 17942, 3872, 3872, 3872, 3872, 0, 14)
c(1, 1, 1, 1, 1, 0.626203386079451, 1, 1, 0.412968905610269, 0.989708974217107, 0.337999336062853, 0.0116189000774594, 0.0775699900409428, 0.0268894544649774, 0, 0.472225296005311, 0.183523293128251, 0.28853601859024, 0.974715060307624, 1, 0.991922098041385, 0.997344251410866, 0.996182361403121, 1, 0.284829036184575, 1, 0.0073033086201173, 0.785769613809893, 0.785769613809893, 0.785769613809893, 0.785769613809893, 1, 0.999225406661503)
c(6, 4, 6, 3, 6, 5, 4, 8, 3, 4, 13, 4, 3, 2, NA, 5, 7, 5, 20, 6, 1, 1, 1, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(10, 14, 9, 6463, 54, 140, 32, 31, 33, 1905, 62, 218, 1556, 286, NA, 61, 16, 103, 167, 219, 1821, 3, 2, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(18074, 13547, 2, 13858, 21, 108, 198, 5, 671, 558, 6, 40, 476, 296, 0, 13, 3, 78, 31, 178, 10052, 197, 196, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(0, 0, 0, 23, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 18261, 18289, NA, NA, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 19872, 19872, NA, NA, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 18351, 18381, NA, NA, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 208, 229, NA, NA, NA, NA, NA, NA, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0.851499391390948, 19.969696969697, 53.8929254475804, 52.2349993234081, 50.5891676618901, 118.994930291508, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0.355605363255071, 11.4618113977747, 2.36410293130679, 1.93168843515652, 2.52120323572528, 53.8831313873642, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0, 2, 48.2881986976457, 45.4169763090071, 37.1908139778481, 1, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, 14, 52.4678771378979, 51.2447941968846, 49.7246742583905, 70, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, 20, 53.5806187824825, 52.3057555608501, 51.110711776471, 124, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, 28, 54.9836024149479, 53.4563836153193, 52.1274912597159, 169, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, 85, 70.7421844688118, 57.5867493801265, 56.2595023940775, 194, NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "▂▁▁▁▇", "▆▇▁▁▁", "▅▇▁▁▁", "▁▂▇▆▁", "▁▁▂▇▂", "▂▅▃▅▇", NA, NA)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1585373897, 1577750400)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1593992053, 1593907200)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1587891533, 1585353600)
c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 13850, 185)

This file contains variables from the \href{https://coronanet-project.org/}{CoronaNet government response project}, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. For more detail on this data, you can see \href{https://docs.google.com/document/d/1zvNMpwj0onFvUZ_gLl4RRjqS-clbHv3TIX6EOHofsME}{the codebook here}.
The format of the data is in country-day-\code{record_id} format. Some
\code{record_id} values have letters appended to indicate that the
general policy category \code{type} also has a value for
\code{type_sub_cat}, which contains more detail about the policy, such
as whether health resources refers to masks, ventilators, or
hospitals. Some entries are marked as \code{new_entry} in the
\code{entry_type} field for when a policy of that type was first
implemented in the country. Later updates to those policies are
marked as updates in \code{entry_type}. To see how policies are
connected, look at the \code{policy_id} field for all policies from
the first entry through updates for a given
country/province/city. If an entry was corrected after initial
data collection, it will read corrected in the \code{entry_type}
field (the original incorrect data has already been replaced with
the corrected data).
}
\references{
\itemize{
\item Robert Kubinec and Luiz Carvalho, "A Retrospective Bayesian Model for Measuring Covariate Effects on Observed COVID-19 Test and Case Counts", \url{https://doi.org/10.31235/osf.io/jp4wk}
}
}
\author{
Kieran Healy
}
\keyword{datasets}
