% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coronanet}
\alias{coronanet}
\title{CoronaNet Government Response Project data}
\format{
A tibble with 18,074 rows and 33 variables:
\describe{
\item{\code{record_id}}{character Unique identifier for each unique policy record}
\item{\code{policy_id}}{double Identifier linking new policies with subsequent updates to policies}
\item{\code{recorded_date}}{date-time When the record was entered into the dataset}
\item{\code{date_announced}}{date When the policy was announced}
\item{\code{date_start}}{date When the policy went into effect}
\item{\code{date_end}}{date When the policy ended (if it has an explicit end date)}
\item{\code{entry_type}}{character Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.}
\item{\code{event_description}}{character A short description of the policy change}
\item{\code{domestic_policy}}{double Indicates where policy targets an area within the initiating country (i.e. is domestic in nature)}
\item{\code{type}}{character The category of the policy}
\item{\code{type_sub_cat}}{character The sub-category of the policy (if one exists)}
\item{\code{type_text}}{double Any additional information about the policy type (such as the number of ventilators/days of quarantine/etc.)}
\item{\code{index_high_est}}{double The high (95\% posterior density) estimate of the country policy activity score (0-100)}
\item{\code{index_med_est}}{double The median (most likely) estimate of the country policy activity score (0-100)}
\item{\code{index_low_est}}{double The low (95\% posterior density) estimate of the country policy activity score (0-100)}
\item{\code{index_country_rank}}{double The relative rank by each day for each country on the policy activity score}
\item{\code{country}}{character The country initiating the policy}
\item{\code{init_country_level}}{character Whether the policy came from the national level or a sub-national unit}
\item{\code{province}}{character Name of sub-national unit}
\item{\code{target_country}}{character Which foreign country a policy is targeted at (i.e. travel policies)}
\item{\code{target_geog_level}}{character Whether the target of the policy is a country as a whole or a sub-national unit of that country}
\item{\code{target_region}}{character The name of a regional grouping (like ASEAN) that is a target of the policy (if any)}
\item{\code{target_province}}{character The name of a province targeted by the policy (if any)}
\item{\code{target_city}}{character The name of a city targeted by the policy (if any)}
\item{\code{target_other}}{logical Any geographical entity that does not fit into the targeted categories coded elsewhere}
\item{\code{target_who_what}}{character Who the policy is targeted at}
\item{\code{target_direction}}{character Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)}
\item{\code{travel_mechanism}}{character If a travel policy, what kind of transportation it affects}
\item{\code{compliance}}{character Whether the policy is voluntary or mandatory}
\item{\code{enforcer}}{character What unit in the country is responsible for enforcement}
\item{\code{link}}{character A link to at least one source for the policy}
\item{\code{iso3}}{character 3-digit ISO country code}
\item{\code{iso2}}{character 2-digit ISO country code}
}
}
\source{
\itemize{
\item Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” BETAVersion 1.0. \url{https://www.coronanet-project.org}.
}
}
\usage{
coronanet
}
\description{
Contains variables from the CoronaNet Government Response Project (\verb{https://coronanet-project.org}), representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. See below for citation information.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab coronanet \cr
   Number of rows \tab 18074 \cr
   Number of columns \tab 33 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 23 \cr
   Date \tab 2 \cr
   numeric \tab 6 \cr
   POSIXct \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   record_id \tab 0 \tab 1.00 \tab 6 \tab 10 \tab 0 \tab 18074 \tab 0 \cr
   policy_id \tab 0 \tab 1.00 \tab 4 \tab 14 \tab 0 \tab 13547 \tab 0 \cr
   entry_type \tab 0 \tab 1.00 \tab 6 \tab 9 \tab 0 \tab 2 \tab 0 \cr
   event_description \tab 0 \tab 1.00 \tab 3 \tab 6463 \tab 0 \tab 13858 \tab 23 \cr
   type \tab 0 \tab 1.00 \tab 6 \tab 54 \tab 0 \tab 21 \tab 0 \cr
   type_sub_cat \tab 6756 \tab 0.63 \tab 5 \tab 140 \tab 0 \tab 108 \tab 0 \cr
   country \tab 0 \tab 1.00 \tab 4 \tab 32 \tab 0 \tab 198 \tab 0 \cr
   init_country_level \tab 0 \tab 1.00 \tab 8 \tab 31 \tab 0 \tab 5 \tab 0 \cr
   province \tab 10610 \tab 0.41 \tab 3 \tab 33 \tab 0 \tab 671 \tab 0 \cr
   target_country \tab 186 \tab 0.99 \tab 4 \tab 1905 \tab 0 \tab 558 \tab 0 \cr
   target_geog_level \tab 11965 \tab 0.34 \tab 13 \tab 62 \tab 0 \tab 6 \tab 0 \cr
   target_region \tab 17864 \tab 0.01 \tab 4 \tab 218 \tab 0 \tab 40 \tab 0 \cr
   target_province \tab 16672 \tab 0.08 \tab 3 \tab 1556 \tab 0 \tab 476 \tab 0 \cr
   target_city \tab 17588 \tab 0.03 \tab 2 \tab 286 \tab 0 \tab 296 \tab 0 \cr
   target_other \tab 18074 \tab 0.00 \tab NA \tab NA \tab 0 \tab 0 \tab 0 \cr
   target_who_what \tab 9539 \tab 0.47 \tab 5 \tab 61 \tab 0 \tab 13 \tab 0 \cr
   target_direction \tab 14757 \tab 0.18 \tab 7 \tab 16 \tab 0 \tab 3 \tab 0 \cr
   travel_mechanism \tab 12859 \tab 0.29 \tab 5 \tab 103 \tab 0 \tab 78 \tab 0 \cr
   compliance \tab 457 \tab 0.97 \tab 20 \tab 167 \tab 0 \tab 31 \tab 0 \cr
   enforcer \tab 0 \tab 1.00 \tab 6 \tab 219 \tab 0 \tab 178 \tab 0 \cr
   link \tab 146 \tab 0.99 \tab 1 \tab 1821 \tab 0 \tab 10052 \tab 0 \cr
   iso3 \tab 48 \tab 1.00 \tab 1 \tab 3 \tab 0 \tab 197 \tab 0 \cr
   iso2 \tab 69 \tab 1.00 \tab 1 \tab 2 \tab 0 \tab 196 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date_start \tab 0 \tab 1.00 \tab 2019-12-31 \tab 2024-05-29 \tab 2020-03-30 \tab 208 \cr
   date_end \tab 12926 \tab 0.28 \tab 2020-01-28 \tab 2024-05-29 \tab 2020-04-29 \tab 229 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   domestic_policy \tab 0 \tab 1.00 \tab 0.85 \tab 0.36 \tab 0.00 \tab 1.00 \tab 1.00 \tab 1.00 \tab 1.00 \tab ▂▁▁▁▇ \cr
   type_text \tab 17942 \tab 0.01 \tab 19.97 \tab 11.46 \tab 2.00 \tab 14.00 \tab 20.00 \tab 28.00 \tab 85.00 \tab ▆▇▁▁▁ \cr
   index_high_est \tab 3872 \tab 0.79 \tab 53.89 \tab 2.36 \tab 48.29 \tab 52.47 \tab 53.58 \tab 54.98 \tab 70.74 \tab ▅▇▁▁▁ \cr
   index_med_est \tab 3872 \tab 0.79 \tab 52.23 \tab 1.93 \tab 45.42 \tab 51.24 \tab 52.31 \tab 53.46 \tab 57.59 \tab ▁▂▇▆▁ \cr
   index_low_est \tab 3872 \tab 0.79 \tab 50.59 \tab 2.52 \tab 37.19 \tab 49.72 \tab 51.11 \tab 52.13 \tab 56.26 \tab ▁▁▂▇▂ \cr
   index_country_rank \tab 3872 \tab 0.79 \tab 118.99 \tab 53.88 \tab 1.00 \tab 70.00 \tab 124.00 \tab 169.00 \tab 194.00 \tab ▂▅▃▅▇ \cr
}


\strong{Variable type: POSIXct}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   recorded_date \tab 0 \tab 1 \tab 2020-03-28 05:38:17 \tab 2020-07-05 23:34:13 \tab 2020-04-26 08:58:53 \tab 13850 \cr
   date_announced \tab 14 \tab 1 \tab 2019-12-31 00:00:00 \tab 2020-07-05 00:00:00 \tab 2020-03-28 00:00:00 \tab 185 \cr
}


This file contains variables from the \href{https://coronanet-project.org/}{CoronaNet government response project}, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. For more detail on this data, you can see \href{https://docs.google.com/document/d/1zvNMpwj0onFvUZ_gLl4RRjqS-clbHv3TIX6EOHofsME}{the codebook here}.
The format of the data is in country-day-\code{record_id} format. Some
\code{record_id} values have letters appended to indicate that the
general policy category \code{type} also has a value for
\code{type_sub_cat}, which contains more detail about the policy, such
as whether health resources refers to masks, ventilators, or
hospitals. Some entries are marked as \code{new_entry} in the
\code{entry_type} field for when a policy of that type was first
implemented in the country. Later updates to those policies are
marked as updates in \code{entry_type}. To see how policies are
connected, look at the \code{policy_id} field for all policies from
the first entry through updates for a given
country/province/city. If an entry was corrected after initial
data collection, it will read corrected in the \code{entry_type}
field (the original incorrect data has already been replaced with
the corrected data).
}
\references{
\itemize{
\item Robert Kubinec and Luiz Carvalho, "A Retrospective Bayesian Model for Measuring Covariate Effects on Observed COVID-19 Test and Case Counts", \url{https://doi.org/10.31235/osf.io/jp4wk}
}
}
\author{
Kieran Healy
}
\keyword{datasets}
