% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coronanet}
\alias{coronanet}
\title{CoronaNet Government Response Project data}
\format{
A tibble with 10,712 rows and 33 variables:
\describe{
\item{\code{record_id}}{character Unique identifier for each unique policy record}
\item{\code{policy_id}}{double Identifier linking new policies with subsequent updates to policies}
\item{\code{recorded_date}}{date-time When the record was entered into the dataset}
\item{\code{date_announced}}{date When the policy was announced}
\item{\code{date_start}}{date When the policy went into effect}
\item{\code{date_end}}{date When the policy ended (if it has an explicit end date)}
\item{\code{entry_type}}{character Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.}
\item{\code{event_description}}{character A short description of the policy change}
\item{\code{domestic_policy}}{double Indicates where policy targets an area within the initiating country (i.e. is domestic in nature)}
\item{\code{type}}{character The category of the policy}
\item{\code{type_sub_cat}}{character The sub-category of the policy (if one exists)}
\item{\code{type_text}}{double Any additional information about the policy type (such as the number of ventilators/days of quarantine/etc.)}
\item{\code{index_high_est}}{double The high (95\% posterior density) estimate of the country policy activity score (0-100)}
\item{\code{index_med_est}}{double The median (most likely) estimate of the country policy activity score (0-100)}
\item{\code{index_low_est}}{double The low (95\% posterior density) estimate of the country policy activity score (0-100)}
\item{\code{index_country_rank}}{double The relative rank by each day for each country on the policy activity score}
\item{\code{country}}{character The country initiating the policy}
\item{\code{init_country_level}}{character Whether the policy came from the national level or a sub-national unit}
\item{\code{province}}{character Name of sub-national unit}
\item{\code{target_country}}{character Which foreign country a policy is targeted at (i.e. travel policies)}
\item{\code{target_geog_level}}{character Whether the target of the policy is a country as a whole or a sub-national unit of that country}
\item{\code{target_region}}{character The name of a regional grouping (like ASEAN) that is a target of the policy (if any)}
\item{\code{target_province}}{character The name of a province targeted by the policy (if any)}
\item{\code{target_city}}{character The name of a city targeted by the policy (if any)}
\item{\code{target_other}}{logical Any geographical entity that does not fit into the targeted categories coded elsewhere}
\item{\code{target_who_what}}{character Who the policy is targeted at}
\item{\code{target_direction}}{character Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)}
\item{\code{travel_mechanism}}{character If a travel policy, what kind of transportation it affects}
\item{\code{compliance}}{character Whether the policy is voluntary or mandatory}
\item{\code{enforcer}}{character What unit in the country is responsible for enforcement}
\item{\code{link}}{character A link to at least one source for the policy}
\item{\code{iso3}}{character 3-digit ISO country code}
\item{\code{iso2}}{character 2-digit ISO country code}
}
}
\source{
\itemize{
\item Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” BETAVersion 1.0. \url{https://www.coronanet-project.org}.
}
}
\usage{
coronanet
}
\description{
Contains variables from the CoronaNet Government Response Project (\verb{https://coronanet-project.org}), representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. See below for citation information.
}
\details{
This file contains variables from the \href{https://coronanet-project.org/}{CoronaNet government response project}, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. For more detail on this data, you can see \href{https://docs.google.com/document/d/1zvNMpwj0onFvUZ_gLl4RRjqS-clbHv3TIX6EOHofsME}{the codebook here}.
The format of the data is in country-day-\code{record_id} format. Some
\code{record_id} values have letters appended to indicate that the
general policy category \code{type} also has a value for
\code{type_sub_cat}, which contains more detail about the policy, such
as whether health resources refers to masks, ventilators, or
hospitals. Some entries are marked as \code{new_entry} in the
\code{entry_type} field for when a policy of that type was first
implemented in the country. Later updates to those policies are
marked as updates in \code{entry_type}. To see how policies are
connected, look at the \code{policy_id} field for all policies from
the first entry through updates for a given
country/province/city. If an entry was corrected after initial
data collection, it will read corrected in the \code{entry_type}
field (the original incorrect data has already been replaced with
the corrected data).
}
\references{
\itemize{
\item Robert Kubinec and Luiz Carvalho, "A Retrospective Bayesian Model for Measuring Covariate Effects on Observed COVID-19 Test and Case Counts", \url{https://doi.org/10.31235/osf.io/jp4wk}
}
}
\author{
Kieran Healy
}
\keyword{datasets}
