% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country Names and ISO codes}
\format{
A data frame with 215 rows and 4 variables:
\describe{
\item{\code{cname}}{character Country name}
\item{\code{iso3}}{character ISO 3 designation}
\item{\code{iso2}}{character ISO 2 designation}
\item{\code{continent}}{Continent}
}
}
\usage{
countries
}
\description{
Convenience table of country names and their abbreviated names
}
\details{
c("character", "character", "character", "character")
c("cname", "iso3", "iso2", "continent")
c(0, 0, 2, 0)
c(1, 1, 0.990697674418605, 1)
c(4, 3, 2, 4)
c(42, 3, 2, 13)
c(0, 0, 0, 0)
c(209, 209, 207, 6)
c(0, 0, 0, 0)

Produced from the ECDC tables in the covdata package.
}
\references{
ISO 2: \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2} ISO 3: \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}
}
\author{
Kieran Healy
}
\keyword{datasets}
