% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat}
\alias{covnat}
\title{International COVID-19 cases and deaths, current as of Monday, February  1, 2021}
\format{
A tibble with 10,210 rows and 8 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{cname}{Name of country (character)}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this week}
\item{deaths}{N reported COVID-19 deaths for this week}
\item{pop}{Country population in 2019}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this week}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this week}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat
}
\description{
A dataset containing weekly national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat) \cr
   Number of rows \tab 10210 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-06 \tab 2021-01-25 \tab 2020-08-10 \tab 56 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   cname \tab 0 \tab 1 \tab 4 \tab 42 \tab 0 \tab 214 \tab 0 \cr
   iso3 \tab 0 \tab 1 \tab 3 \tab 3 \tab 0 \tab 214 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 9767.60 \tab 58732.07 \tab -3864 \tab 8 \tab 184.0 \tab 2530.50 \tab 1782792 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 209.37 \tab 991.05 \tab -875 \tab 0 \tab 2.0 \tab 40.00 \tab 23518 \tab ▇▁▁▁▁ \cr
   pop \tab 15 \tab 1 \tab 40246012.66 \tab 151049839.66 \tab 815 \tab 1293120 \tab 7169456.0 \tab 28515829.00 \tab 1433783692 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1 \tab 142973.31 \tab 864463.00 \tab 0 \tab 179 \tab 2921.5 \tab 37416.25 \tab 25297071 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1 \tab 3873.27 \tab 18988.83 \tab 0 \tab 2 \tab 56.0 \tab 639.00 \tab 421129 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
