% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat}
\alias{covnat}
\title{International COVID-19 cases and deaths, current as of Saturday, January  2, 2021}
\format{
A tibble with 61,623 rows and 8 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{cname}{Name of country (character)}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases on this date}
\item{deaths}{N reported COVID-19 deaths on this date}
\item{pop}{Country population in 2019}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this date}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this date}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat
}
\description{
A dataset containing national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat) \cr
   Number of rows \tab 61623 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-31 \tab 2020-12-13 \tab 2020-07-20 \tab 349 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   cname \tab 0 \tab 1 \tab 4 \tab 42 \tab 0 \tab 213 \tab 0 \cr
   iso3 \tab 0 \tab 1 \tab 3 \tab 3 \tab 0 \tab 213 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 1151.47 \tab 6746.46 \tab -8261 \tab 0 \tab 15 \tab 272.0 \tab 234633 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 26.06 \tab 131.35 \tab -1918 \tab 0 \tab 0 \tab 4.0 \tab 4928 \tab ▁▇▁▁▁ \cr
   pop \tab 58 \tab 1 \tab 41004252.24 \tab 153170670.24 \tab 815 \tab 1293120 \tab 7169456 \tab 28515829.0 \tab 1433783692 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1 \tab 99874.51 \tab 602914.87 \tab 0 \tab 127 \tab 2042 \tab 24393.0 \tab 16067031 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1 \tab 3089.44 \tab 15480.64 \tab 0 \tab 1 \tab 42 \tab 459.5 \tab 297837 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
