% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat}
\alias{covnat}
\title{International COVID-19 cases and deaths to April 9th 2020}
\format{
A tibble with 9,653 rows and 10 columns
\describe{
  \item{date}{date in YYYY-MM-DD format}
  \item{cname}{Name of country (character)}
  \item{iso3}{ISO3 country code (character)}
  \item{cases}{N reported COVID-19 cases on this date}
  \item{deaths}{N reported COVID-19 deaths on this date}
  \item{pop_2018}{Country population in 2018}
  \item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this date}
  \item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this date}
  \item{days_elapsed}{Days elapsed since first reported event (duration)}
  \item{end_label}{Country name for most recent date only, else NA (character)}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat
}
\description{
A dataset containing national-level ECDC data on COVID-19
}
\keyword{datasets}
