% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat}
\alias{covnat}
\title{International COVID-19 cases and deaths, current as of Thursday, February 11, 2021}
\format{
A tibble with 10,638 rows and 8 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{cname}{Name of country (character)}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this week}
\item{deaths}{N reported COVID-19 deaths for this week}
\item{pop}{Country population in 2019}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this week}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this week}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat
}
\description{
A dataset containing weekly national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat) \cr
   Number of rows \tab 10638 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-06 \tab 2021-02-08 \tab 2020-08-17 \tab 58 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   cname \tab 0 \tab 1 \tab 4 \tab 42 \tab 0 \tab 214 \tab 0 \cr
   iso3 \tab 0 \tab 1 \tab 3 \tab 3 \tab 0 \tab 214 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 10008.64 \tab 59146.59 \tab -3864 \tab 9.00 \tab 200 \tab 2708.00 \tab 1782792 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 218.38 \tab 1035.91 \tab -875 \tab 0.00 \tab 2 \tab 43.00 \tab 23518 \tab ▇▁▁▁▁ \cr
   pop \tab 17 \tab 1 \tab 40076152.49 \tab 150633285.14 \tab 815 \tab 1269670.00 \tab 7044639 \tab 28515829.00 \tab 1433783692 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1 \tab 156951.57 \tab 950433.95 \tab 0 \tab 196.25 \tab 3216 \tab 41642.75 \tab 27008565 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1 \tab 4146.04 \tab 20309.91 \tab 0 \tab 2.00 \tab 61 \tab 712.00 \tab 463482 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
