% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat}
\alias{covnat}
\title{International COVID-19 cases and deaths, current as of Sunday, October 11, 2020}
\format{
A tibble with 48,252 rows and 8 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{cname}{Name of country (character)}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases on this date}
\item{deaths}{N reported COVID-19 deaths on this date}
\item{pop}{Country population in 2019}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this date}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this date}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat
}
\description{
A dataset containing national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat) \cr
   Number of rows \tab 48252 \cr
   Number of columns \tab 8 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-31 \tab 2020-10-11 \tab 2020-06-18 \tab 286 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   cname \tab 0 \tab 1 \tab 4 \tab 42 \tab 0 \tab 209 \tab 0 \cr
   iso3 \tab 0 \tab 1 \tab 3 \tab 3 \tab 0 \tab 209 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 772.76 \tab 4660.66 \tab -8261 \tab 0 \tab 11 \tab 171.00 \tab 97894 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 22.25 \tab 126.69 \tab -1918 \tab 0 \tab 0 \tab 3.00 \tab 4928 \tab ▁▇▁▁▁ \cr
   pop \tab 0 \tab 1 \tab 42282049.65 \tab 156394801.85 \tab 815 \tab 1355982 \tab 8082359 \tab 28608715.00 \tab 1433783692 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1 \tab 58376.91 \tab 368550.88 \tab 0 \tab 68 \tab 1172 \tab 12617.75 \tab 7718947 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1 \tab 2240.64 \tab 11876.63 \tab 0 \tab 1 \tab 23 \tab 264.00 \tab 214377 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
