% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat_weekly}
\alias{covnat_weekly}
\title{Weekly International COVID-19 cases and deaths, current as of Wednesday, April 28, 2021}
\format{
A tibble with 12,720 rows and 11 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{year_week}{Year and week of reporting (character, YYYY-WW)}
\item{cname}{Name of country (character)}
\item{pop}{Country population from Eurostat or UN data}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this week}
\item{deaths}{N reported COVID-19 deaths for this week}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this week}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this week}
\item{r14_cases}{14-day notification rate of reported COVID-19 cases per 100,000 population}
\item{r14_deaths}{14-day notification rate of reported COVID-19 cases per 100,000 population}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat_weekly
}
\description{
A dataset containing weekly national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat_wee... \cr
   Number of rows \tab 12720 \cr
   Number of columns \tab 11 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 3 \cr
   numeric \tab 7 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-30 \tab 2021-04-12 \tab 2020-09-14 \tab 68 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   year_week \tab 0 \tab 1 \tab 7 \tab 7 \tab 0 \tab 68 \tab 0 \cr
   cname \tab 0 \tab 1 \tab 4 \tab 33 \tab 0 \tab 213 \tab 0 \cr
   iso3 \tab 0 \tab 1 \tab 3 \tab 3 \tab 0 \tab 213 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pop \tab 0 \tab 1.00 \tab 39883595.79 \tab 150189322.59 \tab 809.00 \tab 1160164.0 \tab 7132530.00 \tab 27691019.00 \tab 1.439324e+09 \tab ▇▁▁▁▁ \cr
   cases \tab 0 \tab 1.00 \tab 11123.87 \tab 60620.94 \tab -28304.00 \tab 13.0 \tab 253.00 \tab 3408.75 \tab 1.782792e+06 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1.00 \tab 237.71 \tab 1114.26 \tab -875.00 \tab 0.0 \tab 3.00 \tab 52.00 \tab 2.351800e+04 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1.00 \tab 227880.18 \tab 1325540.02 \tab 0.00 \tab 318.0 \tab 4996.00 \tab 64319.25 \tab 3.167085e+07 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1.00 \tab 5587.84 \tab 27174.02 \tab 0.00 \tab 4.0 \tab 84.00 \tab 1120.00 \tab 5.672160e+05 \tab ▇▁▁▁▁ \cr
   r14_cases \tab 213 \tab 0.98 \tab 101.09 \tab 219.95 \tab -130.01 \tab 1.2 \tab 11.38 \tab 94.33 \tab 4.345370e+03 \tab ▇▁▁▁▁ \cr
   r14_deaths \tab 213 \tab 0.98 \tab 17.14 \tab 43.33 \tab -67.13 \tab 0.0 \tab 1.14 \tab 13.54 \tab 1.395030e+03 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
