% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat_weekly}
\alias{covnat_weekly}
\title{Weekly International COVID-19 cases and deaths, current as of Wednesday, March  3, 2021}
\format{
A tibble with 11,426 rows and 11 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{year_week}{Year and week of reporting (character, YYYY-WW)}
\item{cname}{Name of country (character)}
\item{pop}{Country population from Eurostat or UN data}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this week}
\item{deaths}{N reported COVID-19 deaths for this week}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this week}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this week}
\item{r14_cases}{14-day notification rate of reported COVID-19 cases per 100,000 population}
\item{r14_deaths}{14-day notification rate of reported COVID-19 cases per 100,000 population}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat_weekly
}
\description{
A dataset containing weekly national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat_wee... \cr
   Number of rows \tab 11426 \cr
   Number of columns \tab 11 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 3 \cr
   numeric \tab 7 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-30 \tab 2021-02-15 \tab 2020-08-17 \tab 60 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   year_week \tab 0 \tab 1.00 \tab 7 \tab 7 \tab 0 \tab 60 \tab 0 \cr
   cname \tab 0 \tab 1.00 \tab 4 \tab 33 \tab 0 \tab 220 \tab 0 \cr
   iso3 \tab 360 \tab 0.97 \tab 3 \tab 3 \tab 0 \tab 214 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pop \tab 0 \tab 1.00 \tab 82089274.77 \tab 378206179.67 \tab 809.00 \tab 1328976.00 \tab 8606033.00 \tab 32365998.00 \tab 4.498460e+09 \tab ▇▁▁▁▁ \cr
   cases \tab 0 \tab 1.00 \tab 21570.12 \tab 129283.62 \tab -3864.00 \tab 10.25 \tab 238.50 \tab 3309.75 \tab 2.699838e+06 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1.00 \tab 481.40 \tab 2684.84 \tab -875.00 \tab 0.00 \tab 3.00 \tab 53.00 \tab 4.827700e+04 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1.00 \tab 357889.39 \tab 2295657.62 \tab 0.00 \tab 239.25 \tab 4026.50 \tab 54322.25 \tab 5.061222e+07 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1.00 \tab 9446.64 \tab 54516.46 \tab 0.00 \tab 3.00 \tab 70.00 \tab 950.75 \tab 1.196316e+06 \tab ▇▁▁▁▁ \cr
   r14_cases \tab 220 \tab 0.98 \tab 88.26 \tab 200.99 \tab -130.01 \tab 1.01 \tab 9.51 \tab 78.66 \tab 4.345370e+03 \tab ▇▁▁▁▁ \cr
   r14_deaths \tab 220 \tab 0.98 \tab 15.70 \tab 41.67 \tab -67.13 \tab 0.00 \tab 0.95 \tab 11.31 \tab 1.395030e+03 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
