% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covnat_weekly}
\alias{covnat_weekly}
\title{Weekly International COVID-19 cases and deaths, current as of Thursday, February 18, 2021}
\format{
A tibble with 11,157 rows and 11 columns
\describe{
\item{date}{date in YYYY-MM-DD format}
\item{year_week}{Year and week of reporting (character, YYYY-WW)}
\item{cname}{Name of country (character)}
\item{pop}{Country population from Eurostat or UN data}
\item{iso3}{ISO3 country code (character)}
\item{cases}{N reported COVID-19 cases for this week}
\item{deaths}{N reported COVID-19 deaths for this week}
\item{cu_cases}{Cumulative N reported COVID-19 cases up to and including this week}
\item{cu_deaths}{Cumulative N reported COVID-19 deaths up to and including this week}
\item{r14_cases}{14-day notification rate of reported COVID-19 cases per 100,000 population}
\item{r14_deaths}{14-day notification rate of reported COVID-19 cases per 100,000 population}
}
}
\source{
\url{http://ecdc.europa.eu/}
}
\usage{
covnat_weekly
}
\description{
A dataset containing weekly national-level ECDC data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab dplyr::ungroup(covnat_wee... \cr
   Number of rows \tab 11157 \cr
   Number of columns \tab 11 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 3 \cr
   Date \tab 1 \cr
   numeric \tab 7 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   year_week \tab 0 \tab 1.00 \tab 7 \tab 7 \tab 0 \tab 59 \tab 0 \cr
   cname \tab 0 \tab 1.00 \tab 4 \tab 33 \tab 0 \tab 219 \tab 0 \cr
   iso3 \tab 354 \tab 0.97 \tab 3 \tab 3 \tab 0 \tab 213 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2019-12-30 \tab 2021-02-08 \tab 2020-08-17 \tab 59 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pop \tab 0 \tab 1.00 \tab 82630200.95 \tab 379479072.41 \tab 809.00 \tab 1318442.00 \tab 8606033.00 \tab 3.2366e+07 \tab 4.498460e+09 \tab ▇▁▁▁▁ \cr
   cases \tab 0 \tab 1.00 \tab 21462.99 \tab 128919.66 \tab -17182.00 \tab 10.00 \tab 231.00 \tab 3.2400e+03 \tab 2.699838e+06 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1.00 \tab 477.63 \tab 2658.06 \tab -875.00 \tab 0.00 \tab 3.00 \tab 5.2000e+01 \tab 4.827700e+04 \tab ▇▁▁▁▁ \cr
   cu_cases \tab 0 \tab 1.00 \tab 344335.08 \tab 2206999.17 \tab 0.00 \tab 222.00 \tab 3792.00 \tab 5.1848e+04 \tab 4.893384e+07 \tab ▇▁▁▁▁ \cr
   cu_deaths \tab 0 \tab 1.00 \tab 9179.44 \tab 52753.35 \tab 0.00 \tab 2.00 \tab 67.00 \tab 9.0900e+02 \tab 1.152349e+06 \tab ▇▁▁▁▁ \cr
   r14_cases \tab 219 \tab 0.98 \tab 86.92 \tab 200.95 \tab -130.01 \tab 0.98 \tab 9.12 \tab 7.5420e+01 \tab 4.345370e+03 \tab ▇▁▁▁▁ \cr
   r14_deaths \tab 219 \tab 0.98 \tab 15.38 \tab 41.64 \tab -67.13 \tab 0.00 \tab 0.90 \tab 1.0680e+01 \tab 1.395030e+03 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
