% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus}
\alias{covus}
\title{COVID-19 data for the USA, current as of Friday, July 17, 2020}
\format{
A tibble with 126,905 rows and 7 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{state}{Two letter State abbreviation (character)}
\item{fips}{State FIPS code (character)}
\item{\code{data_quality_grade}}{character Data quality as assessed by COVID Tracking Project staff}
\item{measure}{Outcome measure for this date}
\item{count}{Count of measure}
\item{\code{measure_label}}{character Outcome measure, suitable for use as a plot label}
}
}
\source{
The COVID-19 Tracking Project \url{https://covidtracking.com} oi
}
\usage{
covus
}
\description{
A dataset containing US state-level data on COVID-19
}
\details{
c("character", "character", "character", "character", "character", "Date", "numeric")
c("state", "fips", "data_quality_grade", "measure", "measure_label", "date", "count")
c(0, 0, 18717, 0, 0, 0, 73451)
c(1, 1, 0.852511721366376, 1, 1, 1, 0.421212718175013)
c(2, 2, 1, 5, 6, NA, NA)
c(2, 2, 2, 24, 54, NA, NA)
c(0, 0, 0, 0, 0, NA, NA)
c(56, 56, 6, 17, 17, NA, NA)
c(0, 0, 0, 0, 0, NA, NA)
c(NA, NA, NA, NA, NA, 18283, NA)
c(NA, NA, NA, NA, NA, 18459, NA)
c(NA, NA, NA, NA, NA, 18393, NA)
c(NA, NA, NA, NA, NA, 177, NA)
c(NA, NA, NA, NA, NA, NA, 62202.1988064504)
c(NA, NA, NA, NA, NA, NA, 261627.828909367)
c(NA, NA, NA, NA, NA, NA, 0)
c(NA, NA, NA, NA, NA, NA, 151)
c(NA, NA, NA, NA, NA, NA, 1294.5)
c(NA, NA, NA, NA, NA, NA, 17204)
c(NA, NA, NA, NA, NA, NA, 5915508)
c(NA, NA, NA, NA, NA, NA, "▇▁▁▁▁")

The measures tracked by the COVID tracking project are as follows:\tabular{ll}{
   measure \tab measure_label \cr
   positive \tab Positive Tests \cr
   negative \tab Negative Tests \cr
   pending \tab Pending Tests \cr
   hospitalized_currently \tab Currently Hospitalized \cr
   hospitalized_cumulative \tab Cumulative Hospitalized \cr
   in_icu_currently \tab Currently in ICU \cr
   in_icu_cumulative \tab Cumulative in ICU \cr
   on_ventilator_currently \tab Currently on Ventilator \cr
   on_ventilator_cumulative \tab Cumulative on Ventilator \cr
   recovered \tab Recovered \cr
   death \tab Deaths \cr
   total_tests_viral \tab Total number of PCR tests performed \cr
   positive_tests_viral \tab Total number of positive PCR tests \cr
   negative_tests_viral \tab Total number of negative PCR tests \cr
   positive_cases_viral \tab Total number of positive cases measured with PCR tests \cr
   death_confirmed \tab Deaths Confirmed \cr
   death_probable \tab Deaths Probable \cr
}


Not all measures are reported by all states.
The \code{positive}, \code{negative}, \code{death}, \code{death_confirmed}, and \code{death_probable} measures are \emph{cumulative} counts.
\code{death_confirmed} is the total number deaths of individuals with COVID-19 infection confirmed by a laboratory test.
In states where the information is available, it tracks only those laboratory-confirmed deaths where COVID also contributed
to the death according to the death certificate. \code{death_probable} is the total number of deaths where COVID was listed as a
cause of death and there is not a laboratory test confirming COVID-19 infection.
}
\keyword{datasets}
