% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus}
\alias{covus}
\title{COVID-19 data for the USA, current as of `r format(Sys.Date(), "%A, %B %e, %Y")`}
\format{
A tibble with `r nrow(covus)` rows and `r ncol(covus)` columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{state}{Two letter State abbreviation (character)}
\item{fips}{State FIPS code (character)}
\item{measure}{Outcome measure for this date}
\item{count}{Count of measure}
\item{pos_neg}{}
\item{death_increase}{}
\item{hospitalized_increase}{}
\item{negative_increase}{}
\item{positive_increase}{}
\item{total_test_results_increase}{}
}
}
\source{
The COVID-19 Tracking Project \url{https://covidtracking.com}
}
\usage{
covus
}
\description{
A dataset containing US state-level data on COVID-19
}
\keyword{datasets}
