% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus}
\alias{covus}
\title{COVID-19 data for the USA, current as of Wednesday, July 22, 2020}
\format{
A tibble with 131,665 rows and 7 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{state}{Two letter State abbreviation (character)}
\item{fips}{State FIPS code (character)}
\item{\code{data_quality_grade}}{character Data quality as assessed by COVID Tracking Project staff}
\item{measure}{Outcome measure for this date}
\item{count}{Count of measure}
\item{\code{measure_label}}{character Outcome measure, suitable for use as a plot label}
}
}
\source{
The COVID-19 Tracking Project \url{https://covidtracking.com}
}
\usage{
covus
}
\description{
A dataset containing US state-level data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus \cr
   Number of rows \tab 131665 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 5 \cr
   Date \tab 1 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1.00 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   fips \tab 0 \tab 1.00 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   data_quality_grade \tab 18717 \tab 0.86 \tab 1 \tab 2 \tab 0 \tab 6 \tab 0 \cr
   measure \tab 0 \tab 1.00 \tab 5 \tab 24 \tab 0 \tab 17 \tab 0 \cr
   measure_label \tab 0 \tab 1.00 \tab 6 \tab 54 \tab 0 \tab 17 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-22 \tab 2020-07-21 \tab 2020-05-13 \tab 182 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   count \tab 75229 \tab 0.43 \tab 67001.62 \tab 284794.9 \tab 0 \tab 152 \tab 1322 \tab 18005 \tab 6536932 \tab ▇▁▁▁▁ \cr
}


The measures tracked by the COVID tracking project are as follows:\tabular{ll}{
   measure \tab measure_label \cr
   positive \tab Positive Tests \cr
   negative \tab Negative Tests \cr
   pending \tab Pending Tests \cr
   hospitalized_currently \tab Currently Hospitalized \cr
   hospitalized_cumulative \tab Cumulative Hospitalized \cr
   in_icu_currently \tab Currently in ICU \cr
   in_icu_cumulative \tab Cumulative in ICU \cr
   on_ventilator_currently \tab Currently on Ventilator \cr
   on_ventilator_cumulative \tab Cumulative on Ventilator \cr
   recovered \tab Recovered \cr
   death \tab Deaths \cr
   total_tests_viral \tab Total number of PCR tests performed \cr
   positive_tests_viral \tab Total number of positive PCR tests \cr
   negative_tests_viral \tab Total number of negative PCR tests \cr
   positive_cases_viral \tab Total number of positive cases measured with PCR tests \cr
   death_confirmed \tab Deaths Confirmed \cr
   death_probable \tab Deaths Probable \cr
}


Not all measures are reported by all states.
The \code{positive}, \code{negative}, \code{death}, \code{death_confirmed}, and \code{death_probable} measures are \emph{cumulative} counts.
\code{death_confirmed} is the total number deaths of individuals with COVID-19 infection confirmed by a laboratory test.
In states where the information is available, it tracks only those laboratory-confirmed deaths where COVID also contributed
to the death according to the death certificate. \code{death_probable} is the total number of deaths where COVID was listed as a
cause of death and there is not a laboratory test confirming COVID-19 infection.
}
\keyword{datasets}
