% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus}
\alias{covus}
\title{COVID-19 data for the USA, current as of Sunday, October 11, 2020}
\format{
A tibble with 372,630 rows and 7 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{state}{Two letter State abbreviation (character)}
\item{fips}{State FIPS code (character)}
\item{\code{data_quality_grade}}{character Data quality as assessed by COVID Tracking Project staff}
\item{measure}{Outcome measure for this date}
\item{count}{Count of measure}
\item{\code{measure_label}}{character Outcome measure, suitable for use as a plot label}
}
}
\source{
The COVID-19 Tracking Project \url{https://covidtracking.com}
}
\usage{
covus
}
\description{
A dataset containing US state-level data on COVID-19
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus \cr
   Number of rows \tab 372630 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-22 \tab 2020-10-11 \tab 2020-06-23 \tab 264 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1.0 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   fips \tab 0 \tab 1.0 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   data_quality_grade \tab 35550 \tab 0.9 \tab 1 \tab 2 \tab 0 \tab 6 \tab 0 \cr
   measure \tab 0 \tab 1.0 \tab 5 \tab 30 \tab 0 \tab 30 \tab 0 \cr
   measure_label \tab 0 \tab 1.0 \tab 6 \tab 54 \tab 0 \tab 30 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   count \tab 249267 \tab 0.33 \tab 183044.8 \tab 730733.3 \tab 0 \tab 284 \tab 3566 \tab 62030 \tab 16047004 \tab ▇▁▁▁▁ \cr
}


The measures tracked by the COVID tracking project are as follows:\tabular{ll}{
   measure \tab measure_label \cr
   positive \tab Positive Tests \cr
   probable_cases \tab Probable Cases \cr
   negative \tab Negative Tests \cr
   pending \tab Pending Tests \cr
   hospitalized_currently \tab Currently Hospitalized \cr
   hospitalized_cumulative \tab Cumulative Hospitalized \cr
   in_icu_currently \tab Currently in ICU \cr
   in_icu_cumulative \tab Cumulative in ICU \cr
   on_ventilator_currently \tab Currently on Ventilator \cr
   on_ventilator_cumulative \tab Cumulative on Ventilator \cr
   recovered \tab Recovered \cr
   death \tab Deaths \cr
   total_tests_viral \tab Total number of PCR tests performed \cr
   positive_tests_viral \tab Total number of positive PCR tests \cr
   negative_tests_viral \tab Total number of negative PCR tests \cr
   positive_cases_viral \tab Total number of positive cases measured with PCR tests \cr
   death_confirmed \tab Deaths Confirmed \cr
   death_probable \tab Deaths Probable \cr
   total_test_encounters_viral \tab Total Test Encounters (PCR) \cr
   total_tests_people_viral \tab Total PCR Tests (People) \cr
   total_tests_antibody \tab Total Antibody Tests \cr
   positive_tests_antibody \tab Positive Antibody Tests \cr
   negative_tests_antibody \tab Negative Antibody Tests \cr
   total_tests_people_antibody \tab Total Antibody Tests (People) \cr
   positive_tests_people_antibody \tab Positive Antibody Tests (People) \cr
   negative_tests_people_antibody \tab Negative Antibody Tests (People) \cr
   total_tests_people_antigen \tab Total Antigen Tests (People) \cr
   positive_tests_people_antigen \tab Positive Antigen Tests (People) \cr
   total_tests_antigen \tab Total Antigen Tests \cr
   positive_tests_antigen \tab Positive Antigen Tests \cr
}


Not all measures are reported by all states.
The \code{positive}, \code{negative}, \code{death}, \code{death_confirmed}, \code{probable_cases} and \code{death_probable} measures are \emph{cumulative} counts.
\code{death_confirmed} is the total number deaths of individuals with COVID-19 infection confirmed by a laboratory test.
In states where the information is available, it tracks only those laboratory-confirmed deaths where COVID also contributed
to the death according to the death certificate. \code{death_probable} is the total number of deaths where COVID was listed as a
cause of death and there is not a laboratory test confirming COVID-19 infection.

For further information on the COVID Tracking Project's measures, see \url{https://covidtracking.com/about-data/data-definitions}
}
\keyword{datasets}
