% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus_ethnicity}
\alias{covus_ethnicity}
\title{COVID-19 case and death counts for the USA by Hispanic/Non-Hispanic ethnicity and state current as of Sunday, December 13, 2020}
\format{
A tibble with 11,547 rows and 7 columns
\describe{
\item{\code{date}}{date Data reported as of this date}
\item{\code{state}}{character State}
\item{\code{group}}{character Ethnic group}
\item{\code{cases}}{integer Total cases, count}
\item{\code{deaths}}{integer Total deaths, count}
\item{\code{hosp}}{integer Total hospitalizations, count}
}
}
\source{
https://covidtracking.com/race
}
\usage{
covus_ethnicity
}
\description{
The COVID Racial Data Tracker advocates for, collects, publishes, and analyzes racial data on the pandemic across the United States.
It’s a collaboration between the COVID Tracking Project and the Boston University Center for Antiracist Research.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus_ethnicity \cr
   Number of rows \tab 11547 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 4 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-04-15 \tab 2020-12-09 \tab 2020-08-12 \tab 69 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   group \tab 0 \tab 1 \tab 7 \tab 12 \tab 0 \tab 3 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 2345 \tab 0.80 \tab 38371.00 \tab 77567.38 \tab 0 \tab 3865.5 \tab 13480.0 \tab 39205 \tab 1227912 \tab ▇▁▁▁▁ \cr
   deaths \tab 2450 \tab 0.79 \tab 1183.28 \tab 2546.89 \tab 0 \tab 50.0 \tab 189.0 \tab 947 \tab 23453 \tab ▇▁▁▁▁ \cr
   hosp \tab 8923 \tab 0.23 \tab 3489.11 \tab 6306.65 \tab 0 \tab 341.0 \tab 1125.5 \tab 3121 \tab 41099 \tab ▇▁▁▁▁ \cr
   tests \tab 10168 \tab 0.12 \tab 326848.55 \tab 719072.10 \tab 0 \tab 13126.5 \tab 83583.0 \tab 305864 \tab 5285250 \tab ▇▁▁▁▁ \cr
}


The \code{group} variable is coded as "Hispanic", "Non-Hispanic", or "Unknown". Hispanics may be of any race. State-level counts should
be handled with care, given the widely varying population distribution of people of different ethnic backgrounds by state.
}
\author{
Kieran Healy
}
\keyword{datasets}
