% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus_race}
\alias{covus_race}
\title{COVID-19 case and death counts for the USA by race and state current as of Wednesday, January  6, 2021}
\format{
A tibble with 38,808 rows and 7 columns
\describe{
\item{\code{date}}{date Data reported as of this date}
\item{\code{state}}{character State}
\item{\code{group}}{character Racial group}
\item{\code{cases}}{integer Total cases, count}
\item{\code{deaths}}{integer Total deaths, count}
\item{\code{hosp}}{integer Total hospitalizations, count}
}
}
\source{
https://covidtracking.com/race
}
\usage{
covus_race
}
\description{
The COVID Racial Data Tracker advocates for, collects, publishes, and analyzes racial data on the pandemic across the United States.
It’s a collaboration between the COVID Tracking Project and the Boston University Center for Antiracist Research.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus_race \cr
   Number of rows \tab 38808 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 4 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-04-12 \tab 2021-01-03 \tab 2020-08-23 \tab 77 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   group \tab 0 \tab 1 \tab 5 \tab 11 \tab 0 \tab 9 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 13122 \tab 0.66 \tab 19719.58 \tab 63202.10 \tab 0 \tab 420.00 \tab 2657.0 \tab 14606.75 \tab 1733813 \tab ▇▁▁▁▁ \cr
   deaths \tab 15034 \tab 0.61 \tab 558.38 \tab 1395.95 \tab 0 \tab 10.00 \tab 57.0 \tab 362.75 \tab 13956 \tab ▇▁▁▁▁ \cr
   hosp \tab 31308 \tab 0.19 \tab 1628.58 \tab 3786.14 \tab -227 \tab 56.00 \tab 282.5 \tab 1456.00 \tab 41099 \tab ▇▁▁▁▁ \cr
   tests \tab 34682 \tab 0.11 \tab 144773.33 \tab 492949.89 \tab 0 \tab 610.75 \tab 10400.5 \tab 84464.50 \tab 6351404 \tab ▇▁▁▁▁ \cr
}


The \code{group} variable is coded as follows:\tabular{l}{
   groups \cr
   White \cr
   Black \cr
   Latino \cr
   Asian \cr
   AI/AN \cr
   NH/PI \cr
   Multiracial \cr
   Other \cr
   Unknown \cr
}


AI/AN is American Indian/Alaska Native. NH/PI is Native Hawaiian/Pacific Islander.
State-level counts should be handled with care, given the widely varying population distribution of people of different racial backgrounds by state.
}
\author{
Kieran Healy
}
\keyword{datasets}
