% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus_race}
\alias{covus_race}
\title{COVID-19 case and death counts for the USA by race and state current as of Tuesday, February  9, 2021}
\format{
A tibble with 43,848 rows and 7 columns
\describe{
\item{\code{date}}{date Data reported as of this date}
\item{\code{state}}{character State}
\item{\code{group}}{character Racial group}
\item{\code{cases}}{integer Total cases, count}
\item{\code{deaths}}{integer Total deaths, count}
\item{\code{hosp}}{integer Total hospitalizations, count}
}
}
\source{
https://covidtracking.com/race
}
\usage{
covus_race
}
\description{
The COVID Racial Data Tracker advocates for, collects, publishes, and analyzes racial data on the pandemic across the United States.
It’s a collaboration between the COVID Tracking Project and the Boston University Center for Antiracist Research.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus_race \cr
   Number of rows \tab 43848 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 4 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-04-12 \tab 2021-02-07 \tab 2020-09-09 \tab 87 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   group \tab 0 \tab 1 \tab 5 \tab 11 \tab 0 \tab 9 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 14543 \tab 0.67 \tab 25575.73 \tab 86033.07 \tab 0 \tab 501 \tab 3218 \tab 18178.0 \tab 2420118 \tab ▇▁▁▁▁ \cr
   deaths \tab 16511 \tab 0.62 \tab 636.54 \tab 1615.41 \tab -1 \tab 11 \tab 63 \tab 403.0 \tab 19624 \tab ▇▁▁▁▁ \cr
   hosp \tab 34557 \tab 0.21 \tab 1872.86 \tab 4251.98 \tab 0 \tab 63 \tab 317 \tab 1595.5 \tab 41099 \tab ▇▁▁▁▁ \cr
   tests \tab 39415 \tab 0.10 \tab 200179.51 \tab 708778.48 \tab 0 \tab 1889 \tab 17467 \tab 120298.0 \tab 15310027 \tab ▇▁▁▁▁ \cr
}


The \code{group} variable is coded as follows:\tabular{l}{
   groups \cr
   White \cr
   Black \cr
   Latino \cr
   Asian \cr
   AI/AN \cr
   NH/PI \cr
   Multiracial \cr
   Other \cr
   Unknown \cr
}


AI/AN is American Indian/Alaska Native. NH/PI is Native Hawaiian/Pacific Islander.
State-level counts should be handled with care, given the widely varying population distribution of people of different racial backgrounds by state.
}
\author{
Kieran Healy
}
\keyword{datasets}
