% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus_race}
\alias{covus_race}
\title{COVID-19 case and death counts for the USA by race and state current as of Wednesday, November 25, 2020}
\format{
A tibble with 32,760 rows and 6 columns
\describe{
\item{\code{date}}{date Data reported as of this date}
\item{\code{state}}{character State}
\item{\code{group}}{character Racial group}
\item{\code{cases}}{integer Total cases, count}
\item{\code{deaths}}{integer Total deaths, count}
\item{\code{hosp}}{integer Total hospitalizations, count}
}
}
\source{
https://covidtracking.com/race
}
\usage{
covus_race
}
\description{
The COVID Racial Data Tracker advocates for, collects, publishes, and analyzes racial data on the pandemic across the United States.
It’s a collaboration between the COVID Tracking Project and the Boston University Center for Antiracist Research.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus_race \cr
   Number of rows \tab 32760 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 3 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-04-12 \tab 2020-11-22 \tab 2020-08-02 \tab 65 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   group \tab 0 \tab 1 \tab 5 \tab 11 \tab 0 \tab 9 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 11361 \tab 0.65 \tab 14431.06 \tab 45600.57 \tab 0 \tab 336.5 \tab 2069 \tab 11058.00 \tab 1040023 \tab ▇▁▁▁▁ \cr
   deaths \tab 13150 \tab 0.60 \tab 490.89 \tab 1238.96 \tab 0 \tab 9.0 \tab 50 \tab 318.00 \tab 10969 \tab ▇▁▁▁▁ \cr
   hosp \tab 27134 \tab 0.17 \tab 1325.53 \tab 3136.86 \tab 0 \tab 49.0 \tab 232 \tab 1172.25 \tab 41099 \tab ▇▁▁▁▁ \cr
}


The \code{group} variable is coded as follows:\tabular{l}{
   groups \cr
   White \cr
   Black \cr
   Latino \cr
   Asian \cr
   AI/AN \cr
   NH/PI \cr
   Multiracial \cr
   Other \cr
   Unknown \cr
}


AI/AN is American Indian/Alaska Native. NH/PI is Native Hawaiian/Pacific Islander.
State-level counts should be handled with care, given the widely varying population distribution of people of different racial backgrounds by state.
}
\author{
Kieran Healy
}
\keyword{datasets}
