% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covus_race}
\alias{covus_race}
\title{COVID-19 case and death counts for the USA by race and state current as of Sunday, December 13, 2020}
\format{
A tibble with 34,641 rows and 7 columns
\describe{
\item{\code{date}}{date Data reported as of this date}
\item{\code{state}}{character State}
\item{\code{group}}{character Racial group}
\item{\code{cases}}{integer Total cases, count}
\item{\code{deaths}}{integer Total deaths, count}
\item{\code{hosp}}{integer Total hospitalizations, count}
}
}
\source{
https://covidtracking.com/race
}
\usage{
covus_race
}
\description{
The COVID Racial Data Tracker advocates for, collects, publishes, and analyzes racial data on the pandemic across the United States.
It’s a collaboration between the COVID Tracking Project and the Boston University Center for Antiracist Research.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab covus_race \cr
   Number of rows \tab 34641 \cr
   Number of columns \tab 7 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 2 \cr
   numeric \tab 4 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-04-15 \tab 2020-12-09 \tab 2020-08-12 \tab 69 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 56 \tab 0 \cr
   group \tab 0 \tab 1 \tab 5 \tab 11 \tab 0 \tab 9 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 11644 \tab 0.66 \tab 16440.79 \tab 51313.03 \tab 0 \tab 375 \tab 2349 \tab 12485.0 \tab 1227912 \tab ▇▁▁▁▁ \cr
   deaths \tab 13429 \tab 0.61 \tab 517.81 \tab 1296.47 \tab 0 \tab 9 \tab 53 \tab 339.0 \tab 11786 \tab ▇▁▁▁▁ \cr
   hosp \tab 28262 \tab 0.18 \tab 1453.98 \tab 3437.65 \tab -227 \tab 51 \tab 254 \tab 1306.5 \tab 41099 \tab ▇▁▁▁▁ \cr
   tests \tab 30876 \tab 0.11 \tab 121291.56 \tab 405697.67 \tab 0 \tab 465 \tab 8409 \tab 71105.0 \tab 5285250 \tab ▇▁▁▁▁ \cr
}


The \code{group} variable is coded as follows:\tabular{l}{
   groups \cr
   White \cr
   Black \cr
   Latino \cr
   Asian \cr
   AI/AN \cr
   NH/PI \cr
   Multiracial \cr
   Other \cr
   Unknown \cr
}


AI/AN is American Indian/Alaska Native. NH/PI is Native Hawaiian/Pacific Islander.
State-level counts should be handled with care, given the widely varying population distribution of people of different racial backgrounds by state.
}
\author{
Kieran Healy
}
\keyword{datasets}
