% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{google_mobility}
\alias{google_mobility}
\title{Google Mobility Data}
\format{
A data frame with 11,453,802 rows and 10 variables:
\describe{
\item{\code{country_region_code}}{character Country Code}
\item{\code{country_region}}{character Country or Region name}
\item{\code{sub_region_1}}{character Subregion (e.g. US state) name}
\item{\code{sub_region_2}}{character Subregion (e.g. US county) name}
\item{\code{metro_area}}{Metropolitan area name}
\item{\code{iso3166_2}}{character ISO 3166-2 Country/Region code}
\item{\code{census_fips_code}}{character US Census FIPS code}
\item{\code{date}}{double Date in yyyy-mm-dd format}
\item{\code{type}}{character Type of location. Values are retail, grocery (and pharmacy), parts, transit (hubs/stations), workplaces, and residential}
\item{\code{pct_diff}}{integer Percent change from baseline activity}
}
}
\source{
Google LLC "Google COVID-19 Community Mobility Reports." https://www.google.com/covid19/mobility/ Accessed: 2020-08-24
}
\usage{
google_mobility
}
\description{
Data from Google's Community Mobility Reports on relative changes in movement trends by location type.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab google_mobility \cr
   Number of rows \tab 11453802 \cr
   Number of columns \tab 10 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 8 \cr
   Date \tab 1 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   country_region_code \tab 8826 \tab 1.00 \tab 2 \tab 2 \tab 0 \tab 134 \tab 0 \cr
   country_region \tab 0 \tab 1.00 \tab 4 \tab 22 \tab 0 \tab 135 \tab 0 \cr
   sub_region_1 \tab 225180 \tab 0.98 \tab 3 \tab 74 \tab 0 \tab 1855 \tab 0 \cr
   sub_region_2 \tab 2245014 \tab 0.80 \tab 2 \tab 50 \tab 0 \tab 8169 \tab 0 \cr
   metro_area \tab 11380902 \tab 0.01 \tab 21 \tab 34 \tab 0 \tab 65 \tab 0 \cr
   iso3166_2 \tab 9221310 \tab 0.19 \tab 4 \tab 6 \tab 0 \tab 2072 \tab 0 \cr
   census_fips_code \tab 8502402 \tab 0.26 \tab 5 \tab 5 \tab 0 \tab 2833 \tab 0 \cr
   type \tab 0 \tab 1.00 \tab 5 \tab 11 \tab 0 \tab 6 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-02-15 \tab 2020-08-21 \tab 2020-05-23 \tab 189 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pct_diff \tab 4509558 \tab 0.61 \tab -14.33 \tab 35.12 \tab -100 \tab -35 \tab -11 \tab 5 \tab 1206 \tab ▇▁▁▁▁ \cr
}


Location accuracy and the understanding of categorized places varies from region to region, so Google does not recommend using this data to compare changes between countries, or between regions with different characteristics (e.g. rural versus urban areas). Regions or categories are omitted if Google does not have have sufficient statistically significant levels of data for it. Changes for each day are compared to a baseline value for that day of the week. The baseline is the median value, for the corresponding day of the week, during the 5-week period Jan 3–Feb 6, 2020. What data is included in the calculation depends on user settings, connectivity, and whether it meets our privacy threshold. If the privacy threshold isn’t met (when somewhere isn’t busy enough to ensure anonymity) we don’t show a change for the day. As a result, you may encounter empty fields for certain places and dates. We calculate these insights based on data from users who have opted-in to Location History for their Google Account, so the data represents a sample of our users. As with all samples, this may or may not represent the exact behavior of a wider population.
}
\author{
Kieran Healy
}
\keyword{datasets}
