% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_sas}
\alias{nchs_sas}
\title{Provisional COVID-19 Death Counts by Sex, Age, and State}
\format{
A tibble with 68,850 rows and 15 variables:
\describe{
\item{\code{data_as_of}}{date Date of data release}
\item{\code{start_date}}{date First date of data period}
\item{\code{end_date}}{date Last date of data period}
\item{\code{group}}{character Unit of time observation: whether data in this row are measured By month, By total, or By year}
\item{\code{year}}{integer Year of observation}
\item{\code{month}}{integer Month of observation}
\item{\code{state}}{character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, \emph{and New York City}, separate from New York state.}
\item{\code{sex}}{character Sex}
\item{\code{age_group}}{character Age group}
\item{\code{covid_19_deaths}}{integer Deaths involving COVID-19 (ICD-code U07.1)}
\item{\code{total_deaths}}{integer Deaths from all causes of death}
\item{\code{pneumonia_deaths}}{integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)}
\item{\code{pneumonia_and_covid_19_deaths}}{integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)}
\item{\code{influenza_deaths}}{integer Influenza Deaths (ICD-10 codes J09-J11)}
\item{\code{pneumonia_influenza_or_covid_19_deaths}}{integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)}
}
}
\source{
National Center for Health Statistics \url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\usage{
nchs_sas
}
\description{
Deaths involving coronavirus disease (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nchs_sas \cr
   Number of rows \tab 68850 \cr
   Number of columns \tab 15 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 6 \cr
   Date \tab 1 \cr
   numeric \tab 8 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   start_date \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 22 \tab 0 \cr
   end_date \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 22 \tab 0 \cr
   group \tab 0 \tab 1 \tab 7 \tab 8 \tab 0 \tab 3 \tab 0 \cr
   state \tab 0 \tab 1 \tab 4 \tab 20 \tab 0 \tab 54 \tab 0 \cr
   sex \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 3 \tab 0 \cr
   age_group \tab 0 \tab 1 \tab 8 \tab 17 \tab 0 \tab 17 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   data_as_of \tab 0 \tab 1 \tab 2021-10-20 \tab 2021-10-20 \tab 2021-10-20 \tab 1 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   year \tab 2754 \tab 0.96 \tab 2020.46 \tab 0.50 \tab 2020 \tab 2020 \tab 2020 \tab 2021 \tab 2021 \tab ▇▁▁▁▇ \cr
   month \tab 8262 \tab 0.88 \tab 6.05 \tab 3.24 \tab 1 \tab 3 \tab 6 \tab 9 \tab 12 \tab ▇▅▅▅▅ \cr
   covid_19_deaths \tab 17426 \tab 0.75 \tab 378.18 \tab 5630.70 \tab 0 \tab 0 \tab 10 \tab 64 \tab 723880 \tab ▇▁▁▁▁ \cr
   total_deaths \tab 10078 \tab 0.85 \tab 2755.89 \tab 42362.22 \tab 0 \tab 42 \tab 154 \tab 683 \tab 5928410 \tab ▇▁▁▁▁ \cr
   pneumonia_deaths \tab 21071 \tab 0.69 \tab 365.54 \tab 5210.91 \tab 0 \tab 0 \tab 18 \tab 83 \tab 649389 \tab ▇▁▁▁▁ \cr
   pneumonia_and_covid_19_deaths \tab 17202 \tab 0.75 \tab 193.10 \tab 2881.39 \tab 0 \tab 0 \tab 0 \tab 31 \tab 369136 \tab ▇▁▁▁▁ \cr
   influenza_deaths \tab 11185 \tab 0.84 \tab 4.16 \tab 78.59 \tab 0 \tab 0 \tab 0 \tab 0 \tab 9372 \tab ▇▁▁▁▁ \cr
   pneumonia_influenza_or_covid_19_deaths \tab 20481 \tab 0.70 \tab 561.74 \tab 8077.07 \tab 0 \tab 0 \tab 26 \tab 122 \tab 1012160 \tab ▇▁▁▁▁ \cr
}


Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis,
and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time
between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes.
This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have
been suppressed in accordance with NHCS confidentiality standards. As of September 2, 2020, this data file includes the following
age groups in addition to the age groups that are routinely included: 0-17, 18-29, 30-49, and 50-64. The new age groups are consistent with categories used across CDC COVID-19 surveillance pages. When analyzing the file, the user should make sure to select only the desired age groups.
Summing across all age categories provided will result in double counting deaths from certain age groups. Similarly, the state variable includes the United States as a whole, and New York City counted separately from the rest of New York State.
The temporal unit of observation also varies, with totals given by year, by month, and overall. It is necessary to first filter the data by desired time unit, region, and age group to ensure there is no double-counting in subsequent calculations.
}
\references{
\url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\author{
Kieran Healy
}
\keyword{datasets}
