% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_sas}
\alias{nchs_sas}
\title{Provisional COVID-19 Death Counts by Sex, Age, and State}
\format{
A tibble with 2,661 rows and 12 variables:
\describe{
\item{\code{data_as_of}}{date Date of data release}
\item{\code{start_week}}{date First week-ending date of data period}
\item{\code{end_week}}{date Last week-ending date of data period}
\item{\code{state}}{character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, \emph{and New York City}, separate from New York state.}
\item{\code{sex}}{character Sex}
\item{\code{age_group}}{character Age group}
\item{\code{covid_19_deaths}}{integer Deaths involving COVID-19 (ICD-code U07.1)}
\item{\code{total_deaths}}{integer Deaths from all causes of death}
\item{\code{pneumonia_deaths}}{integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)}
\item{\code{pneumonia_and_covid_19_deaths}}{integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)}
\item{\code{influenza_deaths}}{integer Influenza Deaths (ICD-10 codes J09-J11)}
\item{\code{pneumonia_influenza_or_covid_19_deaths}}{integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)}
}
}
\source{
National Center for Health Statistics \url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\usage{
nchs_sas
}
\description{
Deaths involving coronavirus disease 2019 (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nchs_sas \cr
   Number of rows \tab 2661 \cr
   Number of columns \tab 12 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 3 \cr
   character \tab 3 \cr
   numeric \tab 6 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   data_as_of \tab 0 \tab 1 \tab 2020-12-02 \tab 2020-12-02 \tab 2020-12-02 \tab 1 \cr
   start_week \tab 0 \tab 1 \tab 2020-02-01 \tab 2020-02-01 \tab 2020-02-01 \tab 1 \cr
   end_week \tab 0 \tab 1 \tab 2020-11-28 \tab 2020-11-28 \tab 2020-11-28 \tab 1 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 4 \tab 20 \tab 0 \tab 54 \tab 0 \cr
   sex \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 4 \tab 0 \cr
   age_group \tab 0 \tab 1 \tab 8 \tab 17 \tab 0 \tab 16 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   covid_19_deaths \tab 390 \tab 0.85 \tab 837.08 \tab 7308.82 \tab 0 \tab 0 \tab 0 \tab 146.50 \tab 249570 \tab ▇▁▁▁▁ \cr
   total_deaths \tab 173 \tab 0.93 \tab 8271.06 \tab 72942.66 \tab 0 \tab 0 \tab 207 \tab 2033.25 \tab 2635214 \tab ▇▁▁▁▁ \cr
   pneumonia_deaths \tab 480 \tab 0.82 \tab 875.03 \tab 7432.91 \tab 0 \tab 0 \tab 12 \tab 196.00 \tab 249148 \tab ▇▁▁▁▁ \cr
   pneumonia_and_covid_19_deaths \tab 369 \tab 0.86 \tab 380.25 \tab 3329.67 \tab 0 \tab 0 \tab 0 \tab 64.00 \tab 113943 \tab ▇▁▁▁▁ \cr
   influenza_deaths \tab 765 \tab 0.71 \tab 27.96 \tab 218.04 \tab 0 \tab 0 \tab 0 \tab 10.00 \tab 6852 \tab ▇▁▁▁▁ \cr
   pneumonia_influenza_or_covid_19_deaths \tab 477 \tab 0.82 \tab 1366.73 \tab 11636.14 \tab 0 \tab 0 \tab 17 \tab 297.25 \tab 390588 \tab ▇▁▁▁▁ \cr
}


Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis,
and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time
between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes.
This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have
been suppressed in accordance with NHCS confidentiality standards.
}
\references{
\url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\author{
Kieran Healy
}
\keyword{datasets}
