% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_sas}
\alias{nchs_sas}
\title{Provisional COVID-19 Death Counts by Sex, Age, and State}
\format{
A tibble with 2,660 rows and 12 variables:
\describe{
\item{\code{data_as_of}}{date Date of data release}
\item{\code{start_week}}{date First week-ending date of data period}
\item{\code{end_week}}{date Last week-ending date of data period}
\item{\code{state}}{character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, \emph{and New York City}, separate from New York state.}
\item{\code{sex}}{character Sex}
\item{\code{age_group}}{character Age group}
\item{\code{covid_19_deaths}}{integer Deaths involving COVID-19 (ICD-code U07.1)}
\item{\code{total_deaths}}{integer Deaths from all causes of death}
\item{\code{pneumonia_deaths}}{integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)}
\item{\code{pneumonia_and_covid_19_deaths}}{integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)}
\item{\code{influenza_deaths}}{integer Influenza Deaths (ICD-10 codes J09-J11)}
\item{\code{pneumonia_influenza_or_covid_19_deaths}}{integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)}
}
}
\source{
National Center for Health Statistics \url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\usage{
nchs_sas
}
\description{
Deaths involving coronavirus disease 2019 (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nchs_sas \cr
   Number of rows \tab 2660 \cr
   Number of columns \tab 12 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 3 \cr
   Date \tab 3 \cr
   numeric \tab 6 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 4 \tab 20 \tab 0 \tab 54 \tab 0 \cr
   sex \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 4 \tab 0 \cr
   age_group \tab 0 \tab 1 \tab 8 \tab 17 \tab 0 \tab 16 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   data_as_of \tab 0 \tab 1 \tab 2020-09-16 \tab 2020-09-16 \tab 2020-09-16 \tab 1 \cr
   start_week \tab 0 \tab 1 \tab 2020-02-01 \tab 2020-02-01 \tab 2020-02-01 \tab 1 \cr
   end_week \tab 0 \tab 1 \tab 2020-09-12 \tab 2020-09-12 \tab 2020-09-12 \tab 1 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   covid_19_deaths \tab 414 \tab 0.84 \tab 618.72 \tab 5377.65 \tab 0 \tab 0 \tab 0 \tab 89.75 \tab 182095 \tab ▇▁▁▁▁ \cr
   total_deaths \tab 175 \tab 0.93 \tab 6149.02 \tab 54237.62 \tab 0 \tab 0 \tab 152 \tab 1503.00 \tab 1957766 \tab ▇▁▁▁▁ \cr
   pneumonia_deaths \tab 486 \tab 0.82 \tab 655.63 \tab 5563.49 \tab 0 \tab 0 \tab 0 \tab 146.75 \tab 186116 \tab ▇▁▁▁▁ \cr
   pneumonia_and_covid_19_deaths \tab 398 \tab 0.85 \tab 276.55 \tab 2407.68 \tab 0 \tab 0 \tab 0 \tab 37.00 \tab 81691 \tab ▇▁▁▁▁ \cr
   influenza_deaths \tab 773 \tab 0.71 \tab 27.39 \tab 213.36 \tab 0 \tab 0 \tab 0 \tab 10.00 \tab 6687 \tab ▇▁▁▁▁ \cr
   pneumonia_influenza_or_covid_19_deaths \tab 492 \tab 0.82 \tab 1031.09 \tab 8746.66 \tab 0 \tab 0 \tab 13 \tab 212.00 \tab 292214 \tab ▇▁▁▁▁ \cr
}


Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis,
and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time
between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes.
This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have
been suppressed in accordance with NHCS confidentiality standards.
}
\references{
\url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\author{
Kieran Healy
}
\keyword{datasets}
