% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_sas}
\alias{nchs_sas}
\title{Provisional COVID-19 Death Counts by Sex, Age, and State}
\format{
A tibble with 46,818 rows and 15 variables:
\describe{
\item{\code{data_as_of}}{date Date of data release}
\item{\code{start_date}}{date First date of data period}
\item{\code{end_date}}{date Last date of data period}
\item{\code{group}}{character Unit of time observation: whether data in this row are measured By month, By total, or By year}
\item{\code{year}}{integer Year of observation}
\item{\code}{month}}{integer Month of observation}
\item{\code{state}}{character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, \emph{and New York City}, separate from New York state.}
\item{\code{sex}}{character Sex}
\item{\code{age_group}}{character Age group}
\item{\code{covid_19_deaths}}{integer Deaths involving COVID-19 (ICD-code U07.1)}
\item{\code{total_deaths}}{integer Deaths from all causes of death}
\item{\code{pneumonia_deaths}}{integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)}
\item{\code{pneumonia_and_covid_19_deaths}}{integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)}
\item{\code{influenza_deaths}}{integer Influenza Deaths (ICD-10 codes J09-J11)}
\item{\code{pneumonia_influenza_or_covid_19_deaths}}{integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)}
}
}
\source{
National Center for Health Statistics \url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\usage{
nchs_sas
}
\description{
Deaths involving coronavirus disease (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nchs_sas \cr
   Number of rows \tab 46818 \cr
   Number of columns \tab 15 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 6 \cr
   Date \tab 1 \cr
   numeric \tab 8 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   start_date \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 14 \tab 0 \cr
   end_date \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 14 \tab 0 \cr
   group \tab 0 \tab 1 \tab 7 \tab 8 \tab 0 \tab 3 \tab 0 \cr
   state \tab 0 \tab 1 \tab 4 \tab 20 \tab 0 \tab 54 \tab 0 \cr
   sex \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 3 \tab 0 \cr
   age_group \tab 0 \tab 1 \tab 8 \tab 17 \tab 0 \tab 17 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   data_as_of \tab 0 \tab 1 \tab 2021-02-24 \tab 2021-02-24 \tab 2021-02-24 \tab 1 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   year \tab 2754 \tab 0.94 \tab 2020.19 \tab 0.39 \tab 2020 \tab 2020 \tab 2020.0 \tab 2020 \tab 2021 \tab ▇▁▁▁▂ \cr
   month \tab 8262 \tab 0.82 \tab 5.79 \tab 3.65 \tab 1 \tab 2 \tab 5.5 \tab 9 \tab 12 \tab ▇▃▃▃▅ \cr
   covid_19_deaths \tab 10769 \tab 0.77 \tab 348.65 \tab 4739.19 \tab 0 \tab 0 \tab 0.0 \tab 58 \tab 478912 \tab ▇▁▁▁▁ \cr
   total_deaths \tab 7105 \tab 0.85 \tab 2558.46 \tab 35837.91 \tab 0 \tab 40 \tab 148.0 \tab 674 \tab 3741316 \tab ▇▁▁▁▁ \cr
   pneumonia_deaths \tab 14087 \tab 0.70 \tab 335.09 \tab 4335.91 \tab 0 \tab 0 \tab 16.0 \tab 81 \tab 414677 \tab ▇▁▁▁▁ \cr
   pneumonia_and_covid_19_deaths \tab 10519 \tab 0.78 \tab 166.78 \tab 2264.61 \tab 0 \tab 0 \tab 0.0 \tab 26 \tab 229982 \tab ▇▁▁▁▁ \cr
   influenza_deaths \tab 9331 \tab 0.80 \tab 6.12 \tab 94.91 \tab 0 \tab 0 \tab 0.0 \tab 0 \tab 8904 \tab ▇▁▁▁▁ \cr
   pneumonia_influenza_or_covid_19_deaths \tab 13820 \tab 0.70 \tab 536.50 \tab 6973.42 \tab 0 \tab 0 \tab 24.0 \tab 123 \tab 671310 \tab ▇▁▁▁▁ \cr
}


Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis,
and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time
between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes.
This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have
been suppressed in accordance with NHCS confidentiality standards. As of September 2, 2020, this data file includes the following
age groups in addition to the age groups that are routinely included: 0-17, 18-29, 30-49, and 50-64.
The new age groups are consistent with categories used across CDC COVID-19 surveillance pages.
When analyzing the file, the user should make sure to select only the desired age groups.
Summing across all age categories provided will result in double counting deaths from certain age groups.
Similarly, the state variable includes the United States as a whole, and New York City counted separately from the rest of New York State.
The temporal unit of observation also varies, with totals given by year, by month, and overall. It is necessary to first filter the data by
desired time unit, region, and age group to ensure there is no double-counting in subsequent calculations.
}
\references{
\url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\author{
Kieran Healy
}
\keyword{datasets}
