% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_sas}
\alias{nchs_sas}
\title{Provisional COVID-19 Death Counts by Sex, Age, and State}
\format{
A tibble with 2,661 rows and 12 variables:
\describe{
\item{\code{data_as_of}}{date Date of data release}
\item{\code{start_week}}{date First week-ending date of data period}
\item{\code{end_week}}{date Last week-ending date of data period}
\item{\code{state}}{character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, \emph{and New York City}, separate from New York state.}
\item{\code{sex}}{character Sex}
\item{\code{age_group}}{character Age group}
\item{\code{covid_19_deaths}}{integer Deaths involving COVID-19 (ICD-code U07.1)}
\item{\code{total_deaths}}{integer Deaths from all causes of death}
\item{\code{pneumonia_deaths}}{integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)}
\item{\code{pneumonia_and_covid_19_deaths}}{integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)}
\item{\code{influenza_deaths}}{integer Influenza Deaths (ICD-10 codes J09-J11)}
\item{\code{pneumonia_influenza_or_covid_19_deaths}}{integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)}
}
}
\source{
National Center for Health Statistics \url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\usage{
nchs_sas
}
\description{
Deaths involving coronavirus disease 2019 (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nchs_sas \cr
   Number of rows \tab 2661 \cr
   Number of columns \tab 12 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 3 \cr
   character \tab 3 \cr
   numeric \tab 6 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   data_as_of \tab 0 \tab 1 \tab 2020-11-12 \tab 2020-11-12 \tab 2020-11-12 \tab 1 \cr
   start_week \tab 0 \tab 1 \tab 2020-02-01 \tab 2020-02-01 \tab 2020-02-01 \tab 1 \cr
   end_week \tab 0 \tab 1 \tab 2020-11-07 \tab 2020-11-07 \tab 2020-11-07 \tab 1 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 4 \tab 20 \tab 0 \tab 54 \tab 0 \cr
   sex \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 4 \tab 0 \cr
   age_group \tab 0 \tab 1 \tab 8 \tab 17 \tab 0 \tab 16 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   covid_19_deaths \tab 394 \tab 0.85 \tab 753.84 \tab 6570.03 \tab 0 \tab 0 \tab 0.0 \tab 123.50 \tab 223984 \tab ▇▁▁▁▁ \cr
   total_deaths \tab 175 \tab 0.93 \tab 7705.50 \tab 67902.99 \tab 0 \tab 0 \tab 192.5 \tab 1898.75 \tab 2452180 \tab ▇▁▁▁▁ \cr
   pneumonia_deaths \tab 481 \tab 0.82 \tab 805.02 \tab 6832.33 \tab 0 \tab 0 \tab 11.0 \tab 183.00 \tab 228934 \tab ▇▁▁▁▁ \cr
   pneumonia_and_covid_19_deaths \tab 373 \tab 0.86 \tab 341.17 \tab 2982.82 \tab 0 \tab 0 \tab 0.0 \tab 52.00 \tab 101910 \tab ▇▁▁▁▁ \cr
   influenza_deaths \tab 770 \tab 0.71 \tab 27.76 \tab 216.32 \tab 0 \tab 0 \tab 0.0 \tab 10.00 \tab 6789 \tab ▇▁▁▁▁ \cr
   pneumonia_influenza_or_covid_19_deaths \tab 485 \tab 0.82 \tab 1254.55 \tab 10650.82 \tab 0 \tab 0 \tab 16.0 \tab 272.00 \tab 356777 \tab ▇▁▁▁▁ \cr
}


Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis,
and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time
between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes.
This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have
been suppressed in accordance with NHCS confidentiality standards.
}
\references{
\url{https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku}
}
\author{
Kieran Healy
}
\keyword{datasets}
