% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_wdc}
\alias{nchs_wdc}
\title{Weekly Counts of Deaths by State and Select Causes 2014-2020}
\format{
A data frame with 256,812 rows and 7 variables:
\describe{
\item{\code{jurisdiction}}{character Jurisdiction of Occurrence}
\item{\code{year}}{double MMWR Year}
\item{\code{week}}{double MMWR Week}
\item{\code{week_ending_date}}{double MMWR Week ending date}
\item{\code{cause_detailed}}{character Cause with ICD Codes}
\item{\code{n}}{double Count of deaths}
\item{\code{cause}}{character Cause of death}
}
}
\source{
2014-2018: \url{https://data.cdc.gov/NCHS/Weekly-Counts-of-Deaths-by-State-and-Select-Causes/3yf8-kanr}. 2019-2020: \url{https://data.cdc.gov/NCHS/Weekly-Counts-of-Deaths-by-State-and-Select-Causes/muzy-jte6}
}
\usage{
nchs_wdc
}
\description{
Final counts of deaths by the week the deaths occurred, by state of occurrence, and by select causes of death for 2014-2018, and Provisional counts of deaths by the week the deaths occurred, by state of occurrence, and by select underlying causes of death for 2019-2020. The dataset also includes weekly provisional counts of death for COVID-19, coded to ICD-10 code U07.1 as an underlying or multiple cause of death.
}
\details{
For 2014-2018, death counts in this dataset were derived from the National Vital Statistics System database that provides the most timely access to the data. Therefore, counts may differ slightly from final data due to differences in processing, recoding, and imputation. For 2019-2020, the dataset also includes weekly provisional counts of death for COVID-19, coded to ICD-10 code U07.1 as an underlying or multiple cause of death.
Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis, and do not represent all deaths that occurred in that period. Data for 2019 and 2020 are provisional and may be incomplete because of the lag in time between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. Causes of death included in this dataset are tabulated by underlying cause of death ICD-10 codes. COVID-19 deaths by underlying cause and multiple cause of death are also included.
}
\references{

}
\author{
Kieran Healy
}
\keyword{datasets}
