% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_wss}
\alias{nchs_wss}
\title{Provisional Death Counts for Coronavirus Disease (COVID-19): Weekly State-Specific Data Updates}
\format{
A tibble with 15,582 rows and 12 variables:
\describe{
\item{\code{data_as_of}}{date Date of analysis}
\item{\code{start_date}}{date Start date of coverage}
\item{\code{end_date}}{date End date of coverage}
\item{\code{year}}{character Year. One of "2020", "2021", or "2020/2021".}
\item{\code{month}}{dbl Month}
\item{\code{obs_unit}}{character Unit of observation. One of: By Total, By Year, By Month.}
\item{\code{state}}{character Geographical unit. One of: the United States, a U.S. State, the District of Columbia, or New York City. New York state measures \emph{do not} include New York City }
\item{\code{race_ethnicity}}{chr Race and ethnic group. One of: Non-Hispanic White, Non-Hispanic Black or African American, Non-Hispanic American Indian or Alaska Native, Non-Hispanic Asian, Non-Hispanic Native Hawaiian or Other Pacific Islander, Non Hispanic more than one race, Hispanic or Latino.}
\item{\code{deaths}}{integer Count of deaths}
\item{\code{dist_pct}}{double Distribution of COVID-19 deaths (\%): Deaths for each group as a percent of the total number of COVID-19 deaths reported.}
\item{\code{uw_dist_pop_pct}}{double Unweighted distribution of population (\%): Population of each group as a percent of the total population.}
\item{\code{wt_dist_pop_pct}}{double Weighted distribution of population (\%): Population of each group as percent of the total population after accounting for how the race and Hispanic origin population is distributed in relation to the geographic areas impacted by COVID-19.}
}
}
\source{
National Center for Health Statistics \url{https://data.cdc.gov/NCHS/Provisional-Death-Counts-for-Coronavirus-Disease-C/pj7m-y5uh}
}
\usage{
nchs_wss
}
\description{
This report provides a weekly summary of deaths with coronavirus disease 2019 (COVID-19) by select geographic and demographic variables.
In this release, counts of deaths are provided by the race and Hispanic origin of the decedent.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nchs_wss \cr
   Number of rows \tab 15582 \cr
   Number of columns \tab 12 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 6 \cr
   numeric \tab 5 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   data_as_of \tab 0 \tab 1 \tab 2023-01-18 \tab 2023-01-18 \tab 2023-01-18 \tab 1 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   start_date \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 37 \tab 0 \cr
   end_date \tab 0 \tab 1 \tab 10 \tab 10 \tab 0 \tab 37 \tab 0 \cr
   year \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 5 \tab 0 \cr
   obs_unit \tab 0 \tab 1 \tab 7 \tab 8 \tab 0 \tab 3 \tab 0 \cr
   state \tab 0 \tab 1 \tab 4 \tab 20 \tab 0 \tab 53 \tab 0 \cr
   race_ethnicity \tab 0 \tab 1 \tab 18 \tab 54 \tab 0 \tab 7 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   month \tab 1855 \tab 0.88 \tab 6.35 \tab 3.52 \tab 1 \tab 3.0 \tab 6.0 \tab 9.0 \tab 12.0 \tab ▇▅▅▅▇ \cr
   deaths \tab 4625 \tab 0.70 \tab 596.40 \tab 8680.87 \tab 0 \tab 0.0 \tab 14.0 \tab 100.0 \tab 718968.0 \tab ▇▁▁▁▁ \cr
   dist_pct \tab 4625 \tab 0.70 \tab 17.59 \tab 29.22 \tab 0 \tab 0.0 \tab 1.1 \tab 19.7 \tab 100.0 \tab ▇▁▁▁▁ \cr
   uw_dist_pop_pct \tab 0 \tab 1.00 \tab 14.28 \tab 23.57 \tab 0 \tab 0.9 \tab 3.1 \tab 12.7 \tab 92.7 \tab ▇▁▁▁▁ \cr
   wt_dist_pop_pct \tab 0 \tab 1.00 \tab 13.68 \tab 21.60 \tab 0 \tab 0.5 \tab 3.2 \tab 14.4 \tab 93.6 \tab ▇▁▁▁▁ \cr
}


The percent of deaths reported in this table are the total number of represent all deaths received and coded as of the date of analysis
and do not represent all deaths that occurred in that period. Data are incomplete because of the lag in time between when the death occurred
and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. \emph{This delay can range from 1 week to 8 weeks or more},
depending on the jurisdiction, age, and cause of death. Provisional counts reported here track approximately 1–2 weeks behind other published data sources
on the number of COVID-19 deaths in the U.S. COVID-19 deaths are defined as having confirmed or presumed COVID-19, and are coded to ICD–10 code U07.1.
Unweighted population percentages are based on the Single-Race Population Estimates from the U.S. Census Bureau, for the year 2018
(available from: https://wonder.cdc.gov/single-race-population.html). Weighted population percentages are computed by multiplying county-level population counts by
the count of COVID deaths for each county, summing to the state-level, and then estimating the percent of the population within each racial and ethnic group.
These weighted population distributions therefore more accurately reflect the geographic locations where COVID outbreaks are occurring.
Jurisdictions are included in this table if more than 100 deaths were received and processed by NCHS as of the data of analysis.

Race and Hispanic-origin categories are based on the 1997 Office of
Management and Budget (OMB) standards (1,2), allowing for the presentation of
data by single race and Hispanic origin. These race and Hispanic-origin
groups—non-Hispanic single-race white, non-Hispanic single-race black or
African American, non-Hispanic single-race American Indian or Alaska Native
(AIAN), non-Hispanic single-race Asian, and non-Hispanic single-race Native
Hawaiian and Other Pacific Islander —differ from the bridged-race categories
shown in most reports using mortality data.

New York State totals exclude New York City (provided in table separately).

Missing values may indicate that a category has between 1 and 9 observed cases and have been suppressed in accordance with NHCS confidentiality standards.
}
\author{
Kieran Healy
}
\keyword{datasets}
