% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nssp_covid_er_nat}
\alias{nssp_covid_er_nat}
\title{NSSP National COVID-related ER Visits}
\format{
A data frame with 54 rows and 9 variables:
\describe{
\item{\code{week}}{integer COLUMN_DESCRIPTION}
\item{\code{num_fac}}{integer COLUMN_DESCRIPTION}
\item{\code{total_ed_visits}}{character COLUMN_DESCRIPTION}
\item{\code{visits}}{integer COLUMN_DESCRIPTION}
\item{\code{pct_visits}}{double COLUMN_DESCRIPTION}
\item{\code{visit_type}}{character COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
\item{\code{source}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{integer COLUMN_DESCRIPTION}
}
}
\source{
Courtesy of Bob Rudis's cdccovidview package
}
\usage{
nssp_covid_er_nat
}
\description{
National Syndromic Surveillance Program (NSSP):
Emergency Department Visits and Percentage of Visits for COVID-19-Like Illness (CLI) or Influenza-like Illness (ILI)
}
\details{
The U.S. Centers for Disease Control provides weekly
summary and interpretation of key indicators that have been adapted
to track the COVID-19 pandemic in the United States. Data is
retrieved using the cdccovidview package from both COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET
(\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}).
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/04102020/nssp-regions.html}
}
\author{
Kieran Healy
}
\keyword{datasets}
