% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nytcovcounty}
\alias{nytcovcounty}
\title{NYT COVID-19 data for US counties, current as of Wednesday, September 16, 2020}
\format{
A tibble with 537,714 rows and 6 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{county}{County name (character)}
\item{state}{State name (character)}
\item{fips}{County FIPS code (character)}
\item{cases}{Cumulative N reported cases}
\item{deaths}{Cumulative N reported deaths}
}
}
\source{
The New York Times \url{https://github.com/nytimes/covid-19-data}
For details on the methods and limitations see \url{https://github.com/nytimes/covid-19-data}.
For county data, note in particular:
\itemize{
\item New York: All cases for the five boroughs of New York City (New York, Kings, Queens, Bronx and Richmond counties) are assigned to a single area called New York City. There is a large jump in the number of deaths on April 6th due to switching from data from New York City to data from New York state for deaths. For all New York state counties, starting on April 8th we are reporting deaths by place of fatality instead of residence of individual.
\item Kansas City, Mo: Four counties (Cass, Clay, Jackson and Platte) overlap the municipality of Kansas City, Mo. The cases and deaths that we show for these four counties are only for the portions exclusive of Kansas City. Cases and deaths for Kansas City are reported as their own line.
\item Alameda County, Calif: Counts for Alameda County include cases and deaths from Berkeley and the Grand Princess cruise ship.
\item Douglas County, Neb. Counts for Douglas County include cases brought to the state from the Diamond Princess cruise ship.
\item Chicago: All cases and deaths for Chicago are reported as part of Cook County.
\item Guam: Counts for Guam include cases reported from the USS Theodore Roosevelt.
}
}
\usage{
nytcovcounty
}
\description{
A dataset containing US county-level data on COVID-19, collected by the New York Times.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nytcovcounty \cr
   Number of rows \tab 537714 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 3 \cr
   Date \tab 1 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   county \tab 0 \tab 1.00 \tab 3 \tab 33 \tab 0 \tab 1923 \tab 0 \cr
   state \tab 0 \tab 1.00 \tab 4 \tab 24 \tab 0 \tab 55 \tab 0 \cr
   fips \tab 5174 \tab 0.99 \tab 5 \tab 5 \tab 0 \tab 3205 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-21 \tab 2020-09-15 \tab 2020-06-24 \tab 239 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 939.45 \tab 5996.93 \tab 0 \tab 14 \tab 73 \tab 354 \tab 255049 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 36.44 \tab 403.12 \tab 0 \tab 0 \tab 1 \tab 9 \tab 23758 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
