% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nytcovcounty}
\alias{nytcovcounty}
\title{NYT COVID-19 data for US counties, current as of Friday, May  1, 2020}
\format{
A tibble with 103,960 rows and 6 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{county}{County name (character)}
\item{state}{State name (character)}
\item{fips}{County FIPS code (character)}
\item{cases}{Cumulative N reported cases}
\item{deaths}{Cumulative N reported deaths}
}
}
\source{
The New York Times \url{https://github.com/nytimes/covid-19-data}
For details on the methods and limitations see \url{https://github.com/nytimes/covid-19-data}.
For county data, note in particular:
\itemize{
\item New York: All cases for the five boroughs of New York City (New York, Kings, Queens, Bronx and Richmond counties) are assigned to a single area called New York City. There is a large jump in the number of deaths on April 6th due to switching from data from New York City to data from New York state for deaths. For all New York state counties, starting on April 8th we are reporting deaths by place of fatality instead of residence of individual.
\item Kansas City, Mo: Four counties (Cass, Clay, Jackson and Platte) overlap the municipality of Kansas City, Mo. The cases and deaths that we show for these four counties are only for the portions exclusive of Kansas City. Cases and deaths for Kansas City are reported as their own line.
\item Alameda County, Calif: Counts for Alameda County include cases and deaths from Berkeley and the Grand Princess cruise ship.
\item Douglas County, Neb. Counts for Douglas County include cases brought to the state from the Diamond Princess cruise ship.
\item Chicago: All cases and deaths for Chicago are reported as part of Cook County.
\item Guam: Counts for Guam include cases reported from the USS Theodore Roosevelt.
}
}
\usage{
nytcovcounty
}
\description{
A dataset containing US county-level data on COVID-19, collected by the New York Times.
}
\keyword{datasets}
