% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nytcovstate}
\alias{nytcovstate}
\title{NYT COVID-19 data for the US states, current as of Monday, August 24, 2020}
\format{
A tibble with 9,584 rows and 5 columns
\describe{
\item{date}{Date in YYYY-MM-DD format (date)}
\item{state}{State name (character)}
\item{fips}{State FIPS code (character)}
\item{cases}{Cumulative N  reported cases}
\item{deaths}{Cumulative N reported deaths}
}
}
\source{
The New York Times \url{https://github.com/nytimes/covid-19-data}.
For details on the methods and limitations see \url{https://github.com/nytimes/covid-19-data}.
}
\usage{
nytcovstate
}
\description{
A dataset containing US state-level data on COVID-19, collected by the New York Times.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nytcovstate \cr
   Number of rows \tab 9584 \cr
   Number of columns \tab 5 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 2 \cr
   Date \tab 1 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 4 \tab 24 \tab 0 \tab 55 \tab 0 \cr
   fips \tab 0 \tab 1 \tab 2 \tab 2 \tab 0 \tab 55 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   date \tab 0 \tab 1 \tab 2020-01-21 \tab 2020-08-23 \tab 2020-05-28 \tab 216 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   cases \tab 0 \tab 1 \tab 37832.45 \tab 78454.76 \tab 1 \tab 1054.75 \tab 8731.5 \tab 40021.25 \tab 669944 \tab ▇▁▁▁▁ \cr
   deaths \tab 0 \tab 1 \tab 1599.11 \tab 4023.03 \tab 0 \tab 18.00 \tab 262.0 \tab 1316.25 \tab 32482 \tab ▇▁▁▁▁ \cr
}
}
\keyword{datasets}
