% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stmf}
\alias{stmf}
\title{Short Term Mortality Fluctuations (STMF) data series}
\format{
A tibble with 302,850 rows and 16 variables:
\describe{
\item{\code{country_code}}{Mortality database country code}
\item{\code{cname}}{character Country name}
\item{\code{iso2}}{character ISO2 country code}
\item{\code{iso3}}{character ISO3 country code}
\item{\code{year}}{double Year}
\item{\code{week}}{double Week number. Each year in the STMF refers to 52 weeks, each week has 7 days. In some cases, the first week of a year may include several days from the previous year or the last week of a year may include days (and, respectively, deaths) of the next year. In particular, it means that a statistical year in the STMF is equal to the statistical year in annual country-specific statistics.}
\item{\code{sex}}{character Sex. m = Males. f = Females. b = Both combined.}
\item{\code{split}}{double Indicates if data were split from aggregated age groups (0 if the original data has necessary detailed age scale). For example, if the original age scale was 0-4, 5-29, 30-65, 65+, then split will be equal to 1}
\item{\code{split_sex}}{double Indicates if the original data are available by sex (0) or data are interpolated (1)}
\item{\code{forecast}}{double Equals 1 for all years where forecasted population exposures were used to calculate weekly death rates.}
\item{\code{approx_date}}{double Approximate date (derived from the \code{year} and \code{week} number).}
\item{\code{age_group}}{character Age group for death counts and rates}
\item{\code{death_count}}{double Weekly death count. This number need not be an integer, because the age categories may be aggregated or split across the source national data.}
\item{\code{death_rate}}{double Weekly death rate.}
\item{\code{deaths_total}}{double Count of deaths for all ages combined.}
\item{\code{rate_total}}{double Crude death rate.}

}
}
\source{
Human Mortality Database, \url{http://mortality.org}
}
\usage{
stmf
}
\description{
Human Mortality Database (HMD) series of weekly death counts across countries.
}
\details{
For further details on the construction of this dataset see the codebook at \url{https://www.mortality.org/Public/STMF_DOC/STMFNote.pdf}.
For the original input data files in standardized form, see \url{https://www.mortality.org/Public/STMF/Inputs/STMFinput.zip}.
}
\references{

}
\author{
Kieran Healy
}
\keyword{datasets}
