% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stmf_country_years}
\alias{stmf_country_years}
\title{Make a table of stmf country years}
\source{
http://
}
\usage{
stmf_country_years(df = stmf)
}
\arguments{
\item{df}{The stmf data frame}
}
\value{
A tibble
}
\description{
Make a table of stmf country years
}
\details{
Get a table of country x year coverage for stmf
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Kieran Healy
}
