% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uspop}
\alias{uspop}
\title{State population estimates for US States}
\format{
A tibble with 459 rows and 17 variables:
\describe{
\item{\code{state}}{character State Name}
\item{\code{state_abbr}}{character State Abbreviation}
\item{\code{statefips}}{character 2-digit FIPS code}
\item{\code{region_name}}{character Census region}
\item{\code{division_name}}{character Census Division}
\item{\code{sex_id}}{character Sex id}
\item{\code{sex}}{character Sex label}
\item{\code{hisp_id}}{character Ethnicity: Hispanic id}
\item{\code{hisp_label}}{character Hispanic label}
\item{\code{fips}}{character Full FIPS code}
\item{\code{pop}}{double Total population}
\item{\code{white}}{double Race alone: White}
\item{\code{black}}{double Race alone: Black or African-American}
\item{\code{amind}}{double Race alone: American Indian and Alaska Native}
\item{\code{asian}}{double Race alone: Asian}
\item{\code{nhopi}}{double Race alone: Native Hawaiian and Other Pacific Islander}
\item{\code{tom}}{double Race alone: Two or more races}
}
}
\source{
https://www.census.gov/data/datasets/time-series/demo/popest/2010s-state-detail.html
}
\usage{
uspop
}
\description{
Population estimates for US States as of July 1st 2018
}
\details{
U.S. Census estimates. Be aware of the US Census classifications of Race and Ethnicity. For the estimated total population for each State, jointly filter on  \code{totsex} in \code{sex_id} and \code{tothisp} in \code{hisp_id}  and then select \code{pop}.
}
\references{
\url{https://www2.census.gov/programs-surveys/popest/tables/2010-2018/state/asrh/PEPSR6H.pdf}
}
\author{
Kieran Healy
}
\keyword{datasets}
