% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uspop}
\alias{uspop}
\title{State population estimates for US States}
\format{
A tibble with 459 rows and 17 variables:
\describe{
\item{\code{state}}{character State Name}
\item{\code{state_abbr}}{character State Abbreviation}
\item{\code{statefips}}{character 2-digit FIPS code}
\item{\code{region_name}}{character Census region}
\item{\code{division_name}}{character Census Division}
\item{\code{sex_id}}{character Sex id}
\item{\code{sex}}{character Sex label}
\item{\code{hisp_id}}{character Ethnicity: Hispanic id}
\item{\code{hisp_label}}{character Hispanic label}
\item{\code{fips}}{character Full FIPS code}
\item{\code{pop}}{double Total population}
\item{\code{white}}{double Race alone: White}
\item{\code{black}}{double Race alone: Black or African-American}
\item{\code{amind}}{double Race alone: American Indian and Alaska Native}
\item{\code{asian}}{double Race alone: Asian}
\item{\code{nhopi}}{double Race alone: Native Hawaiian and Other Pacific Islander}
\item{\code{tom}}{double Race alone: Two or more races}
}
}
\source{
https://www.census.gov/data/datasets/time-series/demo/popest/2010s-state-detail.html
}
\usage{
uspop
}
\description{
Population estimates for US States as of July 1st 2018
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab uspop \cr
   Number of rows \tab 459 \cr
   Number of columns \tab 17 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 10 \cr
   numeric \tab 7 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1.00 \tab 4 \tab 20 \tab 0 \tab 51 \tab 0 \cr
   state_abbr \tab 9 \tab 0.98 \tab 2 \tab 2 \tab 0 \tab 50 \tab 0 \cr
   statefips \tab 0 \tab 1.00 \tab 2 \tab 2 \tab 0 \tab 51 \tab 0 \cr
   region_name \tab 9 \tab 0.98 \tab 4 \tab 9 \tab 0 \tab 4 \tab 0 \cr
   division_name \tab 9 \tab 0.98 \tab 7 \tab 18 \tab 0 \tab 9 \tab 0 \cr
   sex_id \tab 0 \tab 1.00 \tab 4 \tab 6 \tab 0 \tab 3 \tab 0 \cr
   sex \tab 0 \tab 1.00 \tab 4 \tab 10 \tab 0 \tab 3 \tab 0 \cr
   hisp_id \tab 0 \tab 1.00 \tab 4 \tab 7 \tab 0 \tab 3 \tab 0 \cr
   hisp_label \tab 0 \tab 1.00 \tab 5 \tab 12 \tab 0 \tab 3 \tab 0 \cr
   fips \tab 0 \tab 1.00 \tab 11 \tab 11 \tab 0 \tab 51 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pop \tab 0 \tab 1 \tab 2851132.32 \tab 4198641.26 \tab 6154 \tab 386961.5 \tab 1349442 \tab 3558480.0 \tab 39557045 \tab ▇▁▁▁▁ \cr
   white \tab 0 \tab 1 \tab 2179861.40 \tab 3116129.25 \tab 5120 \tab 296294.0 \tab 1088503 \tab 2759335.5 \tab 28531740 \tab ▇▁▁▁▁ \cr
   black \tab 0 \tab 1 \tab 381736.98 \tab 644380.66 \tab 260 \tab 11907.0 \tab 80714 \tab 486281.5 \tab 3673855 \tab ▇▁▁▁▁ \cr
   amind \tab 0 \tab 1 \tab 36143.97 \tab 65036.83 \tab 161 \tab 6103.5 \tab 15273 \tab 35770.5 \tab 651076 \tab ▇▁▁▁▁ \cr
   asian \tab 0 \tab 1 \tab 168458.39 \tab 515557.14 \tab 79 \tab 5045.5 \tab 26484 \tab 140424.5 \tab 6063600 \tab ▇▁▁▁▁ \cr
   nhopi \tab 0 \tab 1 \tab 6966.61 \tab 18657.18 \tab 23 \tab 669.0 \tab 2029 \tab 5063.5 \tab 199872 \tab ▇▁▁▁▁ \cr
   tom \tab 0 \tab 1 \tab 77964.97 \tab 131251.16 \tab 455 \tab 12091.0 \tab 33757 \tab 98669.5 \tab 1554757 \tab ▇▁▁▁▁ \cr
}


U.S. Census estimates. Be aware of the US Census classifications of Race and Ethnicity. For the estimated total population for each State, jointly filter on  \code{totsex} in \code{sex_id} and \code{tothisp} in \code{hisp_id}  and then select \code{pop}.
}
\references{
\url{https://www2.census.gov/programs-surveys/popest/tables/2010-2018/state/asrh/PEPSR6H.pdf}
}
\author{
Kieran Healy
}
\keyword{datasets}
